### 2019-08-29 #1689 @labarba
👋  @kellykochanski
You're missing a comma on this line
https://github.com/kellykochanski/rescal-snow/blob/95b8ac00971f932e8ce2809df5e567b1be068976/docs/paper.bib#L42

Please fix, then run the command `@whedon generate pdf` in a comment below.
