### 2019-08-29 fc1d5f8
Syntax fix in paper.bib
