+++ docs/paper.bib
@@ -39,7 +39,7 @@
 }
 
 @article{Filhol2015,
-  author  = "Filhol, S., and M. Sturm"
+  author  = "Filhol, S., and M. Sturm",
   year    = 2015, 
   title   = "Snow bedforms: A review, new data, and a formation model", 
   journal = "Journal of Geophysical Research: Earth Surface"
@@ -130,7 +130,7 @@
 }
 
 @article{Liston2018,
-  author  = "Liston, G. E., R. B. Haehnel, M. Sturm, C. A. Hiemstra, S. Berezovskaya, R. D. Tabler"
+  author  = "Liston, G. E., R. B. Haehnel, M. Sturm, C. A. Hiemstra, S. Berezovskaya, R. D. Tabler",
   year    = "2007", 
   title   = "Simulating complex snow distributions in windy environments using SnowTran-3D", 
   journal = "Journal of Glaciology", 
