### 2019-08-29 #1689 @kellykochanski
Committing updates to new branch rescal-snow/kk/JOSS-fixes
