### 2019-08-29 #1689 @labarba
@kellykochanski Please go over your bibliography for missing commas.
