### 2019-08-29 #1689 @labarba
👋  @kthyng — Could you handle this submission as editor?
"Rescal-snow: A model for dunes and snow-waves"
