### 2019-08-29 f4fc65e
Syntax fix #2 in paper.bib
