+++ docs/paper.bib
@@ -42,7 +42,7 @@
   author  = "Filhol, S., and M. Sturm",
   year    = 2015, 
   title   = "Snow bedforms: A review, new data, and a formation model", 
-  journal = "Journal of Geophysical Research: Earth Surface"
+  journal = "Journal of Geophysical Research: Earth Surface",
   volume  = 120, 
   issue   = 9, 
   doi     = "10.1002/2015JF003529"
@@ -77,6 +77,7 @@
   pages   = "1267--1281", 
   doi     = "10.5194/tc-13-1267-2019"
 }
+
 @article{Lammel2012,
   author  = "Lammel, M.",
   year    = "2012",
