### 2019-08-29 #1689 @kellykochanski
@labarba Thanks for your help - I should be able to get this. Is there a way to test the pdf compilation outside this issue?
