### 2019-08-29 #1689 @zbeekman
@labarba & @kellykochanski This looks very interesting. I'd be happy to review this with the caveat that I won't be able to start work on this until Sept. 9th.
