### 2019-08-29 #1689 @kthyng
@zbeekman Awesome! That timing is fine, we like to have reviews within about 3 weeks. I'll search for another reviewer now before we start the review process.
