### 2019-08-29 #1689 @kthyng
@rgmyr Are you interested in reviewing this submission for JOSS? You can check out reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
We ask for reviews within about 3 weeks.
