### 2019-08-29 #1689 @kellykochanski
Got it. @labarba one more technical question - the references are not included in the .pdf, even though @whedon was clearly reading them and didn't give me an error. What should I do to trouble shoot?
