### 2019-08-29 #1689 @labarba
I don't know why the references are not rendering. @arfon?

Also, remove the section "Availability," as the JOSS article template includes links to the software repository (and permanent archive) on the margin decorators of the front page.
