### 2019-08-29 1f85e23
Removed Availability from paper.md
