### 2019-08-30 #1689 @arfon
@kellykochanski - this PR should allow your paper to compile: https://github.com/kellykochanski/rescal-snow/pull/25

Also, please make sure the bibtex keys you're citing in your `paper.md` are _exactly_ the same as the ones in the bibtex. You currently have (for example) a citation to `@Kobayashi:1980` in your `paper.md` but the bibtex key in `paper.bib` is `Kobayashi1980`. So you either need to remove the `:` from the string in `paper.md` or change the key in `paper.bib` to `Kobayashi:1980`.

Hopefully that makes sense? It looks like a large number of your references have this issue.
