### 2019-08-30 #1689 @arfon
Just an FYI, the author still needs to fix the citations to their bibtex keys before the references will show up properly.
