### 2019-08-30 2c34d61
Fixing references in paper.md
