### 2019-08-30 40cb23c
Added missing references to paper.bib
