+++ paper/paper.bib
@@ -8,6 +8,38 @@
   doi     = "10.1029/2011WR010745"
 }
 
+@article{Narteau2014,
+  author  = "Rozier, O., and Narteau, C.",
+  year    = 2014,
+  title   = "A real‐space cellular automaton laboratory",
+  journal = "Earth Surface Processes and Landforms",
+  volume  = 39,
+  issue   = 1,
+  pages   = "98-109",
+  doi     = "10.1002/esp.3479"
+}
+
+@article{Narteau2009,
+  author  = "Narteau, C., Zhang, D., Rozier, O., & Claudin, P.",
+  year    = 2009,
+  title   = "Setting the length and time scales of a cellular automaton dune model from the analysis of superimposed bed forms",
+  journal = "Journal of Geophysical Research: Earth Surface",
+  volume  = 114,
+  issue   = F3,
+  doi     = "10.1002/esp.3479"
+}
+
+@article{Liston2007,
+  author  = "Liston, G. E., Haehnel, R. B., Sturm, M., Hiemstra, C. A., Berezovskaya, S., & Tabler, R. D.",
+  year    = 2007,
+  title   = "Simulating complex snow distributions in windy environments using SnowTran-3D",
+  journal = "Journal of Glaciology",
+  volume  = 53,
+  issue   = 181, 
+  pages   = "241-256",
+  doi     = "10.3189/172756507782202865"
+}
+
 @article{Lenaerts2012,
   author  = "Lenaerts, J. T. M., M. R. van den Broeke, W. J. van de Berg, E. van Meijgaard, P. Kuipers Munneke",
   title   = "A new, high-resolution surface mass balance map of Antarctica (1979-2010) based on regional atmospheric climate modeling",
@@ -26,6 +58,17 @@
   pages   = "10342--10348",
   doi     = "10.1002/2015GL066049",
   year    = 2015
+}
+
+@article{Colbeck1983,
+  author  = "Colbeck, S. C.",
+  year    = 1983,
+  title   = "Theory of metamorphism of dry snow",
+  journal = "Journal of Geophysical Research: Oceans",
+  volume  = 88,
+  issue   = C9, 
+  pages   = "5475-5482",
+  doi     = "10.1029/JC088iC09p05475"
 }
 
 @article{Lehning2002,
@@ -109,7 +152,7 @@
   doi     = "10.3189/2013JoG12J094"
 }
 
-@article{Marsh2018,
+@article{2018,
   author  = "Marsh, C. B., R. J. Spiteri, J. W. Pomeroy, and H. S. Wheater",
   year    = "2018",
   journal = "Computers & Geosciences",
@@ -130,7 +173,7 @@
   doi     = "10.1029/2018GL077616"
 }
 
-@article{Liston2018,
+@article{2018,
   author  = "Liston, G. E., R. B. Haehnel, M. Sturm, C. A. Hiemstra, S. Berezovskaya, R. D. Tabler",
   year    = "2007", 
   title   = "Simulating complex snow distributions in windy environments using SnowTran-3D", 
