+++ paper/paper.bib
@@ -77,6 +77,7 @@
   journal = "Cold Regions Science and Technology",
   volume  = 35,
   issue   = 3,
+  year    = 2002,
   pages   = "169--184",
   doi     = "10.1016/S0165-232X(02)00072-1"
 }
@@ -152,7 +153,7 @@
   doi     = "10.3189/2013JoG12J094"
 }
 
-@article{2018,
+@article{Marsh2018,
   author  = "Marsh, C. B., R. J. Spiteri, J. W. Pomeroy, and H. S. Wheater",
   year    = "2018",
   journal = "Computers & Geosciences",
@@ -171,17 +172,6 @@
   issue   = "13", 
   pages   = "6532--6541", 
   doi     = "10.1029/2018GL077616"
-}
-
-@article{2018,
-  author  = "Liston, G. E., R. B. Haehnel, M. Sturm, C. A. Hiemstra, S. Berezovskaya, R. D. Tabler",
-  year    = "2007", 
-  title   = "Simulating complex snow distributions in windy environments using SnowTran-3D", 
-  journal = "Journal of Glaciology", 
-  volume  = "53", 
-  issue   = "181", 
-  pages   = "241--256", 
-  doi     = "10.3189/172756507782202865"
 }
 
 @article{Gallee2012,
