### 2019-08-30 b8d95b7
Removed excess white space around references
