### 2019-08-30 c3f152a
Replacing "&" with "and" in paper.bib
