+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Clark2012,
-  author  = "Clark, M. P., J. Hendrikx, A. G. Slater, D. Kavetski, B. Anderson, N. J. Cullen, T. Kerr, E. O Hreinsson, and R. A. Woods",
+  author  = "Clark, M. P., and J. Hendrikx, and A. G. Slater, and D. Kavetski, and B. Anderson, and N. J. Cullen, and T. Kerr, and E. O Hreinsson, and R. A. Woods",
   year    = 2011,
   title   = "Representing spatial variability of snow water equivalent in hydrologic and land-surface models: A review",
   journal = "Water Resources Research",
@@ -20,7 +20,7 @@
 }
 
 @article{Narteau2009,
-  author  = "Narteau, C., Zhang, D., Rozier, O., & Claudin, P.",
+  author  = "Narteau, C., and Zhang, D., and Rozier, O., and Claudin, P.",
   year    = 2009,
   title   = "Setting the length and time scales of a cellular automaton dune model from the analysis of superimposed bed forms",
   journal = "Journal of Geophysical Research: Earth Surface",
@@ -41,7 +41,7 @@
 }
 
 @article{Lenaerts2012,
-  author  = "Lenaerts, J. T. M., M. R. van den Broeke, W. J. van de Berg, E. van Meijgaard, P. Kuipers Munneke",
+  author  = "Lenaerts, J. T. M. and van den Broeke, M. R. and  van de Berg, W. and van Meijgaard, E., and Munneke, P. K.",
   title   = "A new, high-resolution surface mass balance map of Antarctica (1979-2010) based on regional atmospheric climate modeling",
   journal = "Geophysical Research Letters",
   volume  = 39,
@@ -51,7 +51,7 @@
 }
 
 @article{Blanchard2015,
-  author  = "Blanchard-Wrigglesworth, E., S. L. Farrell, T. Newman, and C. M. Bitz",
+  author  = "Blanchard-Wrigglesworth, and E., S. L. Farrell, and T. Newman, and C. M. Bitz",
   title   = "Snow cover on Arctic sea ice in observations and an Earth System Model",
   journal = "Geophysical Research Letters",
   volume  = 42,
@@ -72,7 +72,7 @@
 }
 
 @article{Lehning2002,
-  author  = "Lehning, M., P. Bartelt, B. Brown, C. Fierz",
+  author  = "Lehning, M. and Bartelt, P. and  Brown, B. and Fierz, C.",
   title   = "A physical SNOWPACK model for the Swiss avalanche warning: Part III: meteorological forcing, thin layer formation and evaluation",
   journal = "Cold Regions Science and Technology",
   volume  = 35,
@@ -92,8 +92,8 @@
   doi     = "10.1002/2015JF003529"
 }
 
-@article{Gallee2012,
-  author  = "Gall\'ee, H., A. Trouvilliez, C. Agosta, C. Genthon, V. Favier, F. Naiim-Bouvet",
+@article{,
+  author  = "Gall\'ee, H., and A. Trouvilliez, and C. Agosta, and C. Genthon, and V. Favier, and F. Naiim-Bouvet",
   year    = 2012, 
   title   = "Transport of snow by the wind: A comparison between observations in Ad'elie Land, Antarctica, and simulations made with the regional climate model MAR", 
   journal = "Boundary-Layer Meteorology", 
@@ -113,7 +113,7 @@
 }
 
 @article{Kochanski2019,
-  author  = "Kochanski, K., R. S. Anderson, and G. E. Tucker",
+  author  = "Kochanski, K., and Anderson, R. S. and Tucker, G. E.",
   year    = 2019, 
[…]
