### 2019-08-30 c3f1eec
Bugfix paper.bib Gallee2012
