+++ paper/paper.bib
@@ -92,7 +92,7 @@
   doi     = "10.1002/2015JF003529"
 }
 
-@article{,
+@article{Gallee2012,
   author  = "Gall\'ee, H., and A. Trouvilliez, and C. Agosta, and C. Genthon, and V. Favier, and F. Naiim-Bouvet",
   year    = 2012, 
   title   = "Transport of snow by the wind: A comparison between observations in Ad'elie Land, Antarctica, and simulations made with the regional climate model MAR", 
