### 2019-08-30 #1689 @kellykochanski
This is now readable. There is still one issue that confuses me: the "Kochanski 2019" reference sometimes renders as "Kochanski, Anderson & Tucker 2019" and sometimes as "Kochanski et al. 2019". This seems slightly arbitrary but does not affect readability.
