+++ paper/paper.bib
@@ -30,7 +30,7 @@
 }
 
 @article{Liston2007,
-  author  = "Liston, G. E., Haehnel, R. B., Sturm, M., Hiemstra, C. A., Berezovskaya, S., & Tabler, R. D.",
+  author  = "Liston, G. E., and Haehnel, R. B., and Sturm, M., and Hiemstra, and C. A., Berezovskaya, S., and Tabler, R. D.",
   year    = 2007,
   title   = "Simulating complex snow distributions in windy environments using SnowTran-3D",
   journal = "Journal of Glaciology",
