### 2019-08-30 003113a
Update paper.md
pdf rendering made it clear that paper appearance would be improved by subheadings
