+++ paper/paper.md
@@ -41,24 +41,29 @@ bibliography: paper.bib
 
 # Summary
 
-One of the most intriguing effects of wind on snow is the creation of mysterious and captivating shapes known as snow bedforms. These features ornament Antarctica, Arctic sea ice, tundra, and many mountain tops [@Filhol2015; @Kochanski2018; @Kobayashi1980].
+When wind blows over dry snow, it creates captivating shapes known as snow bedforms. 
+These features ornament Antarctica, Arctic sea ice, tundra, and many mountain tops [@Filhol2015; @Kochanski2018; @Kobayashi1980].
 They change the reflectivity and average thermal conductivity of snow, and may well change the patterns of snow accumulation and transport.
 They also present a range of interesting self-organized shapes that might interest geomorphologists and Earth surface process scientists.
 Despite these effects, however, they are poorly understood and not yet included in major snow models.
 
+## Requirements of a snow bedform model
 Several recent field studies have identified new computational components needed for a good understanding of snow bedforms.
 First, most snow bedforms (e.g ripples, barchan dunes, snow-steps and sastrugi) are between 0.1 and 2 m in length, with select bedforms (e.g. snow-waves, some whaleback dunes) extending from 5 to 30 m.
 These length scales are based on physical phenomena such as the hop length of blowing snow grains [@Kobayashi1980] and the length scales of turbulent structures in the wind [@Kobayashi1980, @Kochanski2019].
 Existing models of wind-blown snow, however, are designed to model snow redistribution over mountainous [@Liston2007; @Lehning2002; @Marsh2018] or continental [@Gallee2012] scales
 (with grid sizes of 0.1-10 km)
 and are unable to resolve the processes that lead to the formation of snow bedforms.
+
 Second, snow bedforms are the result of a balance between sand-like granular motion and cohesive resistance [@Kochanski2019; @Filhol2015].
 Existing sand dune models (e.g. @Narteau2014, @Lammel2012) successfully model a wide range of granular processes, including sand avalanches and wind-driven saltation, suspension, and creep.
 These models, however, have not had reason to include the kinds of time-dependent cohesion that occur in snow.
-Similarly, the simulations that have advanced our understanding of cohesion and sintering in snow (e.g. @Colbeck1983) have focused on immobile snow, without considering the effects of interrupting metamorphosis with motion.
-Third, snow bedforms tend to move during and immediately after snowfall events [@Kochanski2018], which are also not included in existing dune models.
-These features collectively indicate a need for a new computational model that maps snow processes, such as cohesion and snowfall, into a granular physics framework.
+Similarly, the simulations that have advanced our understanding of cohesion and sintering in snow (e.g. @Colbeck1983, @Lehning2002a) have focused on immobile snow, without considering the effects of interrupting metamorphosis with motion.
 
+Third, snow bedforms tend to move during and immediately after snowfall events [@Kochanski2018], which are also not included in existing dune models.
+Thus, computational studies of snow bedforms will require a model that incorporates snow processes, such as cohesion and snowfall, into a granular physics framework at a relatively high (<0.1 m) resolution.
+
+## Features of Rescal-snow
 Rescal-snow is designed to enable the quantitative study of snow bedforms.
 It simulates 10-100 m domains at 0.05-0.20 m resolution, allowing it to capture all but the smallest snow bedforms.
 The simulation is adapted from a cellular automaton sand dune model, ReSCAL [@Narteau2014], and inherits ReSCAL's granular transport capacities.
@@ -66,24 +71,27 @@ We have added features to simulate processes unique to snow, including snowfall 
[…]
