+++ paper/paper.bib
@@ -173,3 +173,14 @@
   pages   = "6532--6541", 
   doi     = "10.1029/2018GL077616"
 }
+
+@article{Lehning2002a,
+author  = "Lehning, M. and Bartelt, P. and Brown, B. and Fierz, C. and Satyawali, P.",
+year    = 2002,
+title   = "A physical SNOWPACK model for the Swiss avalanche warning: Part II. Snow microstructure",
+journal = "Cold regions science and technology",
+volume  = 35,
+issue   = 3,
+pages   = "147-167",
+doi     = "10.1016/S0165-232X(02)00073-3"
+}
