### 2019-08-30 6cd7466
Cellular automata mechanics image
