### 2019-08-30 6ffaa67
Paper figure size adjusted
