### 2019-08-30 e6b6f52
Revert "Paper figure size adjusted"
