+++ paper/paper.bib
@@ -95,7 +95,7 @@
 @article{Gallee2012,
   author  = "Gall\'ee, H., and A. Trouvilliez, and C. Agosta, and C. Genthon, and V. Favier, and F. Naiim-Bouvet",
   year    = 2012, 
-  title   = "Transport of snow by the wind: A comparison between observations in Ad'elie Land, Antarctica, and simulations made with the regional climate model MAR", 
+  title   = "Transport of snow by the wind: A comparison between observations in Ad\'elie Land, Antarctica, and simulations made with the regional climate model MAR", 
   journal = "Boundary-Layer Meteorology", 
   volume  = 146, 
   issue   = 1, 
