### 2019-09-02 #1699 @whedon
[REVIEW]: Rescal-snow: A model for dunes and snow-waves
**Submitting author:** @kellykochanski (<a href="http://orcid.org/0000-0003-0754-0184">Kelly Kochanski</a>)
**Repository:** <a href="https://github.com/kellykochanski/rescal-snow" target ="_blank">https://github.com/kellykochanski/rescal-snow</a>
**Version:** v1.1.1
