### 2019-09-04 #1699 @kthyng
I just saw that I closed the pre-review issue before whedon had actually added the second reviewer. I have gone in and edited the first comment by hand to add @rgmyr. However, I see that he is not listed as an "Assignee" at the top right of the page, and I can't add him. @arfon is this important? Will @rgmyr have all necessary permissions the way I have added him?
