### 2019-09-04 #1699 @arfon
> @whedon add @rgmyr as reviewer

@kthyng doing this ☝️makes Whedon give the reviewer the permission on the repository (although it doesn't make a new checklist).

@rgmyr will need to accept the invite here: https://github.com/openjournals/joss-reviews/invitations
