### 2019-09-09 #1699 @kthyng
Thanks @arfon! Looks like both reviewers are able to make edits now.
