### 2019-09-10 #1699 @rgmyr
Minor issue on the paper affiliations list: the '2' for LLNL is in the wrong place (comes after instead of before). I think this is because in `paper.md`, that line should just be `index: 2` instead of `- index: 2`.
