### 2019-09-10 kellykochanski/rescal-snow#26 @rgmyr
JOSS Review (@rgmyr): ZLIB not found by CMake
Following the installation notes as written resulted in a CMake error:
```
CMake Error at /usr/share/cmake-3.13/Modules/FindPackageHandleStandardArgs.cmake:137 (message):
  Could NOT find ZLIB (missing: ZLIB_LIBRARY ZLIB_INCLUDE_DIR)
Call Stack (most recent call first):
  /usr/share/cmake-3.13/Modules/FindPackageHandleStandardArgs.cmake:378 (_FPHSA_FAILURE_MESSAGE)
  /usr/share/cmake-3.13/Modules/FindZLIB.cmake:114 (FIND_PACKAGE_HANDLE_STANDARD_ARGS)
  lib/libpng-1.6.37/CMakeLists.txt:42 (find_package)

-- Configuring incomplete, errors occurred!
```
I see that `zlib-1.2.11` is included in `rescal-snow/lib/`, but it doesn't seem to be properly linked by default, at least on my machine. The installation proceeded without that error after manually installing `zlib1g-dev`:
```
$ sudo apt-get install zlib1g-dev
```
