### 2019-09-11 kellykochanski/rescal-snow#27 @rgmyr
JOSS Review (@rgmyr): Release build fails
`make` failed to build after configuring with `-DCMAKE_BUILD_TYPE=Release`. I also tried adding a `-Wno-error` flag to the `make` command, but that didn't seem to have any effect.

I've attached the output: [Release_build_make_output.txt](https://github.com/kellykochanski/rescal-snow/files/3601982/Release_build_make_output.txt). It fails after building/linking `zlib` objects, so maybe this is related to my other issue.

I was able to build after running `cmake` with `-DCMAKE_BUILD_TYPE=Debug`, after which the example scripts seem to be working as expected.
