### 2019-09-11 #1699 @rgmyr
I've had a couple of (possibly related) installation issues, which I've created tickets for:

- [JOSS Review (@rgmyr): ZLIB not found by CMake](https://github.com/kellykochanski/rescal-snow/issues/26)
- [JOSS Review (@rgmyr): Release build fails](https://github.com/kellykochanski/rescal-snow/issues/27)

In the meantime, I now at least have the software running and will try to continue with the rest of the Functionality and Documentation points.
