### 2019-09-17 kellykochanski/rescal-snow#26 @kellykochanski
@r-barnes do you have time to take a look at this?
