### 2019-09-17 kellykochanski/rescal-snow#27 @kellykochanski
@r-barnes Do you have time to take a look at this one?
