+++ docs/rescal-snow-inputs.md
@@ -30,21 +30,21 @@ The run script (run.run, snow_cone.run, etc) is a bash script. You can modify it
 ### The parameter script
 
 This file contains a list of parameters controlling simulation physics.
- - Output_directory : output goes here, useful for data management.
- - Model: rescal-snow can be compiled to allow or disallow certain physics. To ensure the compiled version works with the model you want to run, this must match the compiler flag in src/defs.h. Example: `Model = SNO` here, `#define MODEL_SNO` in src/defs.h.
- - Csp_template : gives a template name and arguments for initial conditions. These are defined and described in src/genesis.c.
- - Phys_prop_file : points to the real data file described in the section below.
- - Simulation size H, L, D : sets size of simulation in cells.
+ - *Output_directory* : output goes here, useful for data management.
+ - *Model*: rescal-snow can be compiled to allow or disallow certain physics. To ensure the compiled version works with the model you want to run, this must match the compiler flag in src/defs.h. Example: `Model = SNO` here, `#define MODEL_SNO` in src/defs.h.
+ - *Csp_template* : gives a template name and arguments for initial conditions. These are defined and described in src/genesis.c.
+ - *Phys_prop_file* : points to the real data file described in the section below.
+ - Simulation size *H*, *L*, *D* : sets size of simulation in cells.
  - Boundary conditions : OPEN, PERIODIC, or REINJECTION. Determines what happens to cells that pass the downwind boundary of the simulation.
- - Lambda_X, Coef_X : control the rates of cell transitions, the main control on model physics. These are best shown graphically: see Fig. 2 of [Narteau et al., 2009](dx.doi.org/10.1029/2008JF001127).
+ - *Lambda_X*, *Coef_X* : control the rates of cell transitions, which define/control the model physics. These are best shown graphically: see Fig. 2 of [Narteau et al., 2009](dx.doi.org/10.1029/2008JF001127).
 
 The parameters you are likely to want to modify for snow studies include:
- - Csp_template : the initial conditions, see src/genesis.c
- - Tau_min : the control for wind speed - note that wind speed *decreases* non-linearly as you increase Tau_min
- - Lambda_I : the snowfall rate (must use `Csp_template = SNOWFALL`)
- - Lambda_S : the sintering rate
- - Lambda_F : the hardness of sintered snow, relative to the hardness of freshly-fallen snow
- - Coef_A : the ratio between vertical and horizontal transport of snow in air; controls the particle settling speed
+ - *Csp_template* : the initial conditions, see src/genesis.c
+ - *Tau_min* : the control for wind speed - note that wind speed *decreases* non-linearly as you increase Tau_min
+ - *Lambda_I* : the snowfall rate (must use `Csp_template = SNOWFALL`)
+ - *Lambda_S* : the sintering rate
+ - *Lambda_F* : the hardness of sintered snow, relative to the hardness of freshly-fallen snow
+ - *Coef_A* : the ratio between vertical and horizontal transport of snow in air; controls the particle settling speed
 
 ### The real_data folder
 
@@ -52,12 +52,12 @@ All of the information we have given rescal-snow so far has been non-dimensional
 This file contains information about the physical situation that rescal-snow represents. For more discussion on converting rescal-snow output to reality, see 'docs/calibration_and_validation.md'.
 
 The file scripts/real_data/sealevel_snow.prop contains the following variables:
- - kappa: the von Karman constant, dimensionless, generally agreed to be near 0.4
- - z_0: the [aerodynamic roughness length](https://en.wikipedia.org/wiki/Roughness_length) of the surface, m
- - rhoair: the density of the fluid that moves the grains, kg/m^3
[…]
