### 2019-09-17 kellykochanski/rescal-snow#28 @kellykochanski
Fixed small suggestion: italicized parameter and variable names in `docs/rescal-snow-inputs.md`
