### 2019-09-17 kellykochanski/rescal-snow#28 @rgmyr
@kellykochanski Thanks! Also I realize that some of the above points may seem nitpicky. I'm open to the argument that this is not primarily a python package, and I will be lenient as far as submission acceptance goes (although I do stand by them as suggestions, and hope that they're helpful in improving your Python work generally). The most important things to me are that docstrings use triple quotes, be placed under the `def` and `class` lines, and that there be some overview of `utilities` file contents.
