### 2019-09-17 kellykochanski/rescal-snow#28 @kellykochanski
@rgmyr I appreciate the suggestions. `rescal_utilities` has been developed by several people with different coding backgrounds and styles. During development, I focused on having a consistent level of functional documentation and examples, but did not set standards for any of the points you have mentioned above.

Most of these should be relatively straightforward changes.
