### 2019-09-17 kellykochanski/rescal-snow#28 @kellykochanski
As a follow-up, however, this release is not a python package; the python components are tools for simulations setup, data management, and analysis, which we expect will be useful to some but not all users.
