### 2019-09-17 2ad77c8
Replaced "from numpy import *" in csp_creator
Replaced zeros with np.zeros, int8 with np.int8
