### 2019-09-17 kellykochanski/rescal-snow#28 @kellykochanski
@21dmohan @defaziogiancarlo You should take a look at these comments - I'm addressing them for Rescal-snow, but a lot of @rgmyr's feedback is relevant to your functions and will help you polish your future Python code.
