### 2019-09-17 9570a8d
Improving doc consistency in rescal_utilities.py
Moved docstrings below function name, defined all with triple quotes """
Changed class names to CamelCase
Replaced double underscores (name mangling) with single underscores (non-mangled non-public names)
