+++ scripts/utilities/rescal_utilities.py
@@ -2,7 +2,7 @@
 # Utilities to set up run scripts quickly and easily; these tools are designed to aid parameter space
 # explorations, sensitivity analyses, and large batches of runs.
 # Parameters : Class to hold, update, change, or write all the parameters that ReSCAL needs to run
-# Run_Script : Class to hold, update, change, or write a ReSCAL run script with appropriate flags
+# RunScript : Class to hold, update, change, or write a ReSCAL run script with appropriate flags
 # This includes almost all options and inputs, except those in the "real_data" file.
 
 import os
@@ -11,35 +11,38 @@ import itertools
 import subprocess
 import glob
 
-class Design_a_run():
-    # Design a ReSCAL run
-    # This interfaces with both the Parameters and Run_Script utilities
-    # in order to produce run script and parameter files simultaneously
-    # using a standardized input format
+class DesignRun():
+    """
+    Design a ReSCAL run
+    This interfaces with both the Parameters and RunScript utilities
+    in order to produce run script and parameter files simultaneously
+    using a standardized input format
+    """
 
     def __init__(self):
         self.name       = 'run'
         self.directory  = '.'
         self.parameters = Parameters()
-        self.run_script = Run_Script()
+        self.run_script = RunScript()
         self.parameters.set_header("Test from write_tests.py")
 
     def set_parameters(self, param_dict):
-        # Sets all parameters, given as {name : value} pairs
-        # Allows user to mix things that go in parameter file
-        # with things that go in the run script, etc
-        # TODO make it possible to do compiler flags too?
+        """
+        Sets all parameters, given as {name : value} pairs
+        Allows user to mix things that go in parameter file
+        with things that go in the run script, etc
+        """
 
         for name in param_dict.keys():
-            if self.__is_a_parameter(name):
+            if self._is_a_parameter(name):
                 self.parameters.set({name : param_dict[name]})
-            elif self.__is_a_run_script_option(name):
+            elif self._is_a_run_script_option(name):
                 self.run_script.set({name : param_dict[name]})
             else:
                 print("Warning : skipped nonexistent parameter " + name)
 
     def list_all(self):
-        # Lists all available options/parameters
+        """Lists all available options/parameters"""
         return self.run_script.list_all(), self.parameters.list_all()
 
     def get(self, param):
@@ -60,11 +63,11 @@ class Design_a_run():
     def set_directory(self, directory):
         self.directory = directory
 
-    def __is_a_parameter(self, name):
+    def _is_a_parameter(self, name):
         # Checks whether the parameters dictionary contains 'name'
         return ((name == 'Environment') or (name in self.parameters.list_all()))
 
-    def __is_a_run_script_option(self, name):
+    def _is_a_run_script_option(self, name):
         # Checks whether 'name' is an option in the run script
         return (name in self.run_script.list_all())
 
@@ -77,38 +80,44 @@ class Design_a_run():
 ##--------------------------------------------------------------------------------------
 
 class Parameters():
-        # Holds all the parameters that ReSCAL needs to run
+    """Holds all the parameters that ReSCAL needs to run"""
     def __init__(self):
         self.parameters                 = {}
         self.parameter_descriptions     = {}
-        self.__special_parameter_descriptions = {}
-        self.__set_special =                    {}
-        self.__default_header()
-        self.__default_parameters()
+        self._special_parameter_descriptions = {}
+        self._set_special =                    {}
+        self._default_header()
+        self._default_parameters()
 
     def new_parameter(self, name, description, default_value):
-        # New parameter
-        # Name should be a string (the parameter name expected by ReSCAL)
-        # Description should be a string describing its use or purpose etc
-        # Default value is the default value, generally a string or float
+        """
+        New parameter
[…]
