### 2019-09-17 kellykochanski/rescal-snow#28 @21dmohan
Thank you for the feedback! I will definitely utilize this information and resources as I continue to code!
