### 2019-09-17 kellykochanski/rescal-snow#26 @kellykochanski
This error also occurs on my new machine prior to installing zlib.
