### 2019-09-18 6ec85fc
Standardized docstrings in scripts/utilities
