+++ scripts/utilities/cellspace.py
@@ -1,11 +1,11 @@
-__doc__ = '''Utilities to read, write, modify, and visualize rescal-snow cell spaces.'''
+__doc__ = """Utilities to read, write, modify, and visualize rescal-snow cell spaces."""
 __author__ = 'Gian-Carlo DeFazio'
 __date__ = 'August 16 2019'
 
-import matplotlib
 import sys
 import os
-# Matplotlib will fail if no display is available (e.g. many high-performance computing environments)
+import matplotlib
+# Matplotlib will fail if no display is available (as in many high-performance computing environments)
 if bool(os.environ.get('DISPLAY', None)) == False:
 	matplotlib.use('Agg')
 import matplotlib.pyplot as plt
@@ -14,7 +14,6 @@ import numpy as np
 import random
 import gzip
 import scipy.ndimage
-import random
 
 import heightmap
 
@@ -52,9 +51,11 @@ cell_types = {
 
 
 def in_bounds(base, overlay, position):
-    '''Evaluates if overlay can be placed onto base at 
+    """
+    Evaluates if overlay can be placed onto base at 
     position. position is the coordinates of base on which the top left 
-    corner of overlay is placed.'''
+    corner of overlay is placed.
+    """
     # verify that the dimensions match
     if base.ndim != len(position):
         return False
@@ -81,7 +82,9 @@ def in_bounds(base, overlay, position):
 
 
 def non_border(a):
-    '''Get a slice of a that excludes border cells.'''
+    """
+    Get a slice of a 1D array that excludes border cells.
+    """
     # the stuff we care about has states 0,1,2,3
     valid = np.nonzero(a < 4)[0]
     return a[valid[0]: valid[-1]+1]
@@ -89,7 +92,7 @@ def non_border(a):
 
 # replaced with find_air_or_mobile
 def surface_position(a):
-    '''find index of first solid sand starting from the top.'''
+    """find index of first solid sand starting from the top."""
     sand_indices = np.nonzero(a == 0)[0]
     if sand_indices.size > 0:
         return sand_indices[0]
@@ -112,8 +115,10 @@ def find_air_or_mobile(column):
 
 
 def shift_fill(a, shift, fill=0):
-    '''Shift an array to the right by shift and fill in with fill value.
-    If shift value is negative, the shift will be to the left.'''
+    """
+    Shift an array to the right by shift and fill in with fill value.
+    If shift value is negative, the shift will be to the left.
+    """
     # nothing to do
     if shift == 0:
         pass
@@ -130,9 +135,11 @@ def shift_fill(a, shift, fill=0):
     
 
 def change_surface_level(column, delta):
-    '''Given a column in the 3D cellspace, modify the sand height by delta.
+    """
+    Given a column in the 3D cellspace, modify the sand height by delta.
     Clip at the boundaries of the space. Try to keep mobile sand by moving it with
-    the surface.'''
+    the surface.
+    """
 
     # if nothing to do, do nothing
     if delta == 0:
@@ -164,13 +171,15 @@ def change_surface_level(column, delta):
 
 
 class CellSpace:
-    '''A snow-rescal cell space representation. This class deals with the .csp files
+    """
+    A snow-rescal cell space representation. This class deals with the .csp files
     made by rescal and genesis. The .csp files contain a header followed by the 3D space
-    of cells as a 1D array.'''
+    of cells as a 1D array.
+    """
 
-    
     def __init__(self, filename):
-        '''read in a .csp or .csp.gz file and create a CellSpace instance
+        """
+        read in a .csp or .csp.gz file and create a CellSpace instance
         to represent the file contents. filename can also be a bytearray object
         that contains the contents of a .csp file.
 
@@ -193,7 +202,8 @@ class CellSpace:
 
[…]
