+++ scripts/utilities/heightmap.py
@@ -1,22 +1,19 @@
-__doc__ = '''Utilities to read and visualize rescal-snow height maps.'''
+__doc__ = """Utilities to read and visualize rescal-snow height maps."""
 __author__ = 'Gian-Carlo DeFazio'
 __date__ = '16 August 2019'
 
+import sys
+import os
 import math
 import numpy as np
 import scipy.ndimage
 import matplotlib
-import sys
-import os
 # Matplotlib will fail if no display is available (e.g. many high-performance computing environments)
 if bool(os.environ.get('DISPLAY', None)) == False:
 	matplotlib.use('Agg')
-    
 import matplotlib.pyplot as plt
 from matplotlib import cm
 from mpl_toolkits.mplot3d import Axes3D
-
-
 
 # space invader height map
 invader_template = np.array([[0,0,1,0,0,0,0,0,1,0,0],
@@ -30,10 +27,18 @@ invader_template = np.array([[0,0,1,0,0,0,0,0,1,0,0],
 
 
 def gaussian_hill(amplitude, sigma, height_padding, width_padding):
-    '''Creates a guassian heightmap. The center of the guassian
+    """
+    Creates a guassian heightmap. The center of the guassian
     will have a value of amplitude. sigma can be either a scaler 
     or a 2-tuple. The padding determines the size of the array returned.
-    The size is (2*height_padding+1, 2*width_padding+1)'''
+    The size is (2*height_padding+1, 2*width_padding+1)
+
+    Inputs:
+     - amplitude: height in cells of gaussian hills at center (highest point)
+     - sigma: width (at standard deviation) of hills. 
+          may be scalar (symmetric hills) or 2-tuple (asymmetric)
+     - height_padding, width_padding: determine size of array returned
+     """
     
     # does it the cheap way:
     # makes an impulse, meaning a 1 in the middle and 0 otherwise,
@@ -48,13 +53,22 @@ def gaussian_hill(amplitude, sigma, height_padding, width_padding):
 
 
 def make_sinusoid(height, frequency, dims, phase=0, wind_direction=True, no_negative=False):
-    '''Create a sinusoid on a 2D plane that has dimensions dims. 
+    """
+    Create a sinusoid on a 2D plane that has dimensions dims. 
     Does an inverse Fourier transform and discretizes. 
     The wave is resized so that its height ranges from 0 to height. 
     The frequency is relative to the space. If the frequency is x, wave will 
     complete x cycles across the space. If wind_direction is True, then the wave
     heights will vary in the horizontal direction, otherwise they will vary in the
-    vertical direction.'''
+    vertical direction.
+
+    Inputs:
+     - height: wave ranges from 0 to height
+     - dims: dimensions of array
+     - wind_direction: boolean, if True wave varies in horizontal, if False in vertical
+     - no_negative: disallow negative heights, returning zeros as minimum
+
+    """
     # create a 2D array of the same size as the dims
     grid = np.zeros(dims, dtype=np.complex)
 
@@ -81,17 +95,27 @@ def make_sinusoid(height, frequency, dims, phase=0, wind_direction=True, no_nega
 
 
 def scale(height_map, amplitude=1, vertical=1, horizontal=1):
-    '''Scales a height map. The heigh map dimensions are scaled in the 
+    """
+    Scales a height map. The heigh map dimensions are scaled in the 
     horizontal and vertical directions. The values are then scaled by a factor
-    of amplitude.'''
+    of amplitude.
+
+    Inputs:
+     - height_map
+     - amplitude: scale factor for stretching height map
+     - vertical, horizontal: relative scale factors in horizontal and vertical dimensions
+    """
     return np.kron(height_map, np.full((vertical, horizontal), amplitude))
 
 
 def fft2d_analyze(data):
[…]

[…]
