### 2019-09-18 0faf1e4
Standardized docstrings in scripts/utilities
