+++ scripts/utilities/README.md
@@ -1,0 +1,99 @@
+# Rescal-snow utilites
+This folder contains scripts for configuring, running and analyzing Rescal-snow simulations. These are run through Python3 and bash. 
+They make the Rescal-snow workflow more consistent and reproducible by making it possible to do most of the work in Python and store both input and output in Python classes.
+
+## [rescal_utilities.py](rescal_utilities.py)
+Utilities to set up Rescal-snow runs quickly and easily.
+These tools are designed to aid parameter space explorations, sensitivity analyses, and large batches of runs.
+The following classes manage most Rescal-snow options and inputs (excepting those in the "real_data" file):
+
+  Parameters : Class to hold, update, change, read, or write all the parameters that ReSCAL needs to run
+  
+  RunScript  : Class to hold, update, change, read, or write a ReSCAL run script with appropriate flags
+  
+  DesignRun  : Umbrella class to hold, update, change, read or write all parameters, sorting them into Parameters and RunScript
+
+Example usage:
+```python
+$python
+
+# Create the DesignRun object and (optional but advised) a basic description
+my_run     = rescal_utilities.DesignRun()
+my_run.set_name("my_run")escal.sbatch
+(base) kelly@Beauty:~/Downloads/rescal-snow/scripts/utilities$ vim example_pyrescal
+(base) kelly@Beauty:~/Downloads/rescal-snow/scripts/utilities$ vim example
+my_run.set_header("An example run varying L, D and Lambda_S, and looking for the rescal executable in build")
+
+# Describe values for some subset of Rescal-snow parameterescal.sbatch
+(base) kelly@Beauty:~/Downloads/rescal-snow/scripts/utilities$ vim example_pyrescal
+(base) kelly@Beauty:~/Downloads/rescal-snow/scripts/utilities$ vim examples
+# Others will take defaults from rescal_utilities.Parameters._default_parameters() and rescal_utilities.RunScript._default_options()
+parameters = {'L':500, 'D':50, 'Lambda_S':0.01, 'rescallocation':'../../build'}
+my_run.set_parameters(parameters)
+for the output should be created. The default is RESCAL_SNOW_ROOT/data_runs.
+# Create a run script named my_run.run and a parameter script named my_run.par
+my_run.write()
+```
+See also the usage in [scripts/utilities/param_space_exploration_example.py](scripts/utilities/param_space_exploration_example.py)
+
+## Example: [param_space_exploration_example.py](param_space_exploration_example.py)
+This is an example file showing how to run a parameter space exploration. In this example, most simulation parameters are set to fixed values, then 30 runs are created to vary:
+  five snowfall rates, controlled by parameter Lambda_I,
+  six wind speeds, controlled by parameter Tau_min,
+  and all 30 combinations thereof.
+ The script creates 30 new directories, one for each run. Each directory is seeded with a run script, a parameter file, and rescal and genesis executables. This allows all 30 simulations to run in parallel withou
+t interferance.
+
+## [dataruns.py](dataruns.py)
+A python interface to Rescal-snow. This sets Rescal-snow parameters for one or more runs, runs the simulation, and keeps the results in pythonic arrays for ongoing processing.
+A DataRun object takes in the parameters and meta-parameters required to run Rescal-snow. The DataRun can receive and processthe output of Rescal-snow while Rescal-snow is running.
[…]
