### 2019-09-18 f4f5a6c
Typo fix utilities/README
