### 2019-09-18 6af1f26
Increased docstring coverage in utilities
