### 2019-09-18 40dcef7
Increased docstring coverage in utilities
