+++ scripts/utilities/rescal_utilities.py
@@ -1,4 +1,4 @@
-___author__ = "Kelly Kochanski"
+
 __date__    = "Jun 25 2018"
 __doc__     = r"""Utilities to set up run scripts quickly and easily; these tools are designed to aid parameter space
  explorations, sensitivity analyses, and large batches of runs.
@@ -8,10 +8,6 @@ __doc__     = r"""Utilities to set up run scripts quickly and easily; these tool
 """
 
 import os
-import re
-import itertools
-import subprocess
-import glob
 
 class DesignRun():
     """
@@ -48,6 +44,7 @@ class DesignRun():
         return self.run_script.list_all(), self.parameters.list_all()
 
     def get(self, param):
+        """get a parameter value from Parameters() object"""
         if param in self.parameters.list_all():
             return self.parameters.get(param)
         elif param in self.run_script.list_all():
@@ -118,8 +115,8 @@ class Parameters():
     def _default_parameters(self):
         """
         Default values for parameters
-        Initialized here for sahara sand
-        Basic model parameters and geometry of model run:
+        Used by all runs setup with the DesignRun workflow unless overwritten with the set_parameters() function.
+        Default values are sahara sand.
         """
         self.new_parameter('Model', 'Defines a set of transitions for grains in the cellular automaton e.g. DUN or SNO', 'SNO')
         self.new_parameter('Output_directory', 'Directory where output files go', './out')
@@ -348,7 +345,7 @@ class RunScript():
 
     def _set_flag_option(self, descriptive_name, flag, value):
         """
-        Some options are command line flags. Some are not.
+        Some options are command line flags. Some are set elsewhere in the run script.
         Try to hide this bit of extra complexity from the user.
         """
         self._flag_options[descriptive_name]   = flag
@@ -462,196 +459,3 @@ class RunScript():
         # make file executable
         os.chmod(filename, 0o764)
 
-
-##--------------------------------------------------------------------------------------
-##--------------------------------------------------------------------------------------
-##--------------------------------------------------------------------------------------
-
-
-
-# TODO allow iteratables to be used
-# not just lists
-# TODO allow passing of arguments in dicts, but with no guaranteed ordering
-def all_parameter_combos(named_parameter_lists):
-    """
-    Takes a parameter list of lists of lists such as:
-       [['Coef_A', [0.1, 0.2, 0.3]], ['random seed', [1,2,3]]])
-       so that each sublist has the variable name and set of parameters to vary over.
-       Creates a generator that returns a dictionary with each possible parameter
-       combination and a string that can be used to identify the varying parameters.
-       If variable names have spaces they are replaced with underscores.
-    """
-
-    # separate parameter names from values lists, cast all values to strings
-    names = [x[0] for x in named_parameter_lists]
-    parameter_lists_original = [x[1] for x in named_parameter_lists]
-
-    # cast all values to strings
-    parameter_lists = [[str(i) for i in x] for x in parameter_lists_original]
-
-    # create a generator that makes all combination of parameters lists (the cartesian product)
-    # also creates a string for the file name
-    for parameter_combination in itertools.product(*parameter_lists):
-        # create a dict of the names and some value combination
-        params_dict_to_add = dict(zip(names, parameter_combination))
-        # creats a directory name based on the names and values
-        # if any of the values are file paths, only the base name of the file is used
-        directory_name_suffix = [name.replace(' ', '_') + '-' + \
-                                 os.path.basename(str(params_dict_to_add[name])) for name in names]
[…]
