### 2019-09-18 kellykochanski/rescal-snow#28 @kellykochanski
@rgmyr I believe I've addressed your points in the JOSS-fixes branch:

- Added a README for scripts/utilities
- (Related: re-organized files to make their purposes easier to define and simplify dependencies)
- More descriptive __doc__ attributes at the top of each file
- Added a handful of missing docstrings (left one-line "set" and "get" etc functions without)
- All docstrings now use triple quotes """ with consistent line breaks for multiline strings
- Inputs and outputs broken out onto separate lines in appropriate docstrings
- Classes now CamelCase, methods now under_scored
- double underscores generally replaced with single underscores to avoid unneeded name mangling
- instances of `import *` excised
