### 2019-09-18 kellykochanski/rescal-snow#27 @rgmyr
@r-barnes I'm on [Pop!_OS 19.04](https://distrowatch.com/table.php?distribution=popos), which is Ubuntu-based. My home machine is vanilla Ubuntu 16.04, so I can also check tonight and see if I get the same issue there.
