### 2019-09-18 kellykochanski/rescal-snow#28 @rgmyr
@kellykochanski Awesome, looks good to me. I will close this issue and mark `Functionality documentation` completed.
