### 2019-09-18 kellykochanski/rescal-snow#27 @kellykochanski
@r-barnes I reproduced this bug, and the associated zlib issue, on Ubuntu Budgie.
