### 2019-09-18 kellykochanski/rescal-snow#26 @kellykochanski
@egreen77 I know it's been a while, but could you fill @r-barnes (who has made additional changes to the rescal build system) in on the work you did on the Rescal build system and the zlib dependency?
