### 2019-09-18 29e1180
Typo in paper.md
Fixed affiliation indexing syntax
