### 2019-09-18 kellykochanski/rescal-snow#26 @r-barnes
I don't think there's much light to be shed on this from that direction: I've replaced zlib with an upgraded version and replaced the build system with cmake, so there's not a lot of original structure left.
