### 2019-09-18 kellykochanski/rescal-snow#26 @r-barnes
Since cmake has superior package management, the simplest solution here may be to drop both zlib and libpng from the codebase and request the user install these manually. cmake can enforce package versions that are known to work. @kellykochanski?

I'll keep checking in how to fix this using the existing setup, but haven't found solutions yet.
