### 2019-09-18 8091090
Fixed libpng not finding zlib dependency/
