### 2019-09-18 f87753b
Fixed unused result warnings
