+++ src/lgca.c
@@ -566,7 +566,10 @@ void lecture_mvt() {
     exit(-4);
   }
 
-  fread(CelMvt, sizeof(MvtField), CHLD, fp);
+  if (fread(CelMvt, sizeof(MvtField), CHLD, fp) != sizeof(MvtField) * CHLD) {
+    ErrPrintf("Read error\n");
+    exit(-1);
+  }
 
   fclose(fp);
 
