### 2019-09-18 kellykochanski/rescal-snow#27 @kellykochanski
@egreen77 I now get the following build error:

```bash
make[2]: *** No rule to make target '../lib/zlib-1.2.11/libz.a', needed by 'bin2png'.  Stop.
make[2]: *** Waiting for unfinished jobs....
[ 94%] Building C object CMakeFiles/bin2png.dir/src/bin2png.c.o
make[2]: *** No rule to make target '../lib/zlib-1.2.11/libz.a', needed by 'csp2png'.  Stop.
make[2]: *** Waiting for unfinished jobs....
[ 94%] Building C object CMakeFiles/csp2png.dir/src/csp2png.c.o
[ 95%] Building C object CMakeFiles/rescal.dir/src/drand48.c.o
[ 96%] Building C object CMakeFiles/rescal.dir/src/entry.c.o
make[2]: *** No rule to make target '../lib/zlib-1.2.11/libz.a', needed by 'rescal'.  Stop.
make[2]: *** Waiting for unfinished jobs....
[ 96%] Building C object CMakeFiles/rescal.dir/src/rescal.c.o
make[1]: *** [CMakeFiles/Makefile2:73: CMakeFiles/bin2png.dir/all] Error 2
make[1]: *** [CMakeFiles/Makefile2:295: CMakeFiles/rescal.dir/all] Error 2
make[1]: *** [CMakeFiles/Makefile2:147: CMakeFiles/csp2png.dir/all] Error 2
make: *** [Makefile:130: all] Error 2
```
