### 2019-09-18 kellykochanski/rescal-snow#27 @r-barnes
@rgmyr: Do you know your compiler version? Most compilers are getting progressively better at detecting warning conditions, so we'd prefer to resolve those whenever possible rather than using `-Wno-error`.
