### 2019-09-19 4ffdea5
Drop ExternalProject (extraneous) and HOMEPAGE_URL (needless restriction on cmake version)
