### 2019-09-19 af26f5c
Add ExternalProject back (used now)
