+++ CMakeLists.txt
@@ -1,5 +1,7 @@
 cmake_minimum_required (VERSION 3.9)
 
+include(ExternalProject)
+
 project (rescal-snow
   DESCRIPTION "Simulating snow self-organization with cellular automata"
   LANGUAGES C
