### 2019-09-19 8aff6d5
Fixed unused result warnings (#31)
Fixes issues with `fread` results not being checked.
