### 2019-09-19 681229d
Fix issue with fread checking the wrong number of elements
