### 2019-09-19 737393e
Added info about installing libpng and zlib to README
