### 2019-09-19 2b802f1
Cast fread to int32_t
