### 2019-09-19 9f6c218
Use find_package instead of including png and zlib ourselves (#30)
* Use find_package instead of including png and zlib ourselves

* Fix issue with fread checking the wrong number of elements

* Added info about installing libpng and zlib to README

* Cast fread to int32_t
