### 2019-09-19 kellykochanski/rescal-snow#26 @kellykochanski
@rgmyr We've resolved this issue (and the related issue about building with `cmake -Wno-dev -DCMAKE_BUILD_TYPE=Release`).

Libpng and zlib are now explicit dependencies and are not packaged with the software.
