### 2019-09-19 7ce6173
Removing references to obselete how_to_install doc
