### 2019-09-23 kellykochanski/rescal-snow#26 @r-barnes
Closing this for now.
