### 2019-09-24 #1699 @kellykochanski
@rgmyr We've updated the build system and installation instructions to fix these issues. Zlib is now a dependency and is not explicitly linked within Rescal-snow.
