### 2019-09-25 kellykochanski/rescal-snow#27 @rgmyr
Just for reference, my `gcc --version` is `8.3.0`. Release build now works after installing `libpng-dev` and `zlib1g-dev`.
