### 2019-09-25 #1699 @rgmyr
My installation issues were fixed by the build update and additional dependency instructions.

My only other change request was regarding documentation and style consistency of the included Python utilities, which @kellykochanski promptly improved:
- [JOSS Review (@rgmyr): Python utilities documentation and style](https://github.com/kellykochanski/rescal-snow/issues/28)

I should note that I wasn't able to test the functionality related to parallel simulation runs, simply because I don't have access to any remote computing clusters at this time. With that said, it looks pretty straightforward and reasonable based on my previous experiences with `slurm`-scheduled systems.

That completes my review checklist, so I'm happy to recommend acceptance of this submission.
