### 2019-09-26 e9a2a26
JOSS-Paper: Remove ambiguous/subjective adjective
