+++ paper/paper.md
@@ -45,7 +45,7 @@ bibliography: paper.bib
 
 # Summary
 
-When wind blows over dry snow, it creates captivating shapes known as snow bedforms. 
+When wind blows over dry snow, it creates a range of unique shapes known as snow bedforms.
 These features, which include snow dunes, waves, snow-steps and sastrugi (figure below), ornament Antarctica, Arctic sea ice, tundra, and mountain ridges [@Filhol2015; @Kochanski2018; @Kobayashi1980].
 They change the reflectivity and average thermal conductivity of snow, and may change the patterns of snow accumulation and transport.
 Despite these effects, however, they are poorly understood and their effects are not yet included in major snow or climate models.
