### 2019-09-26 7afecbb
JOSS-paper: I think this needs a citation.
  If you mention "several recent studies" then you should probably cite
  a few of them.
