### 2019-09-26 1971967
JOSS-Paper: Reference & clarify purpose of figure
