### 2019-09-26 63cbd6b
JOSS-Paper: Fix typos
  Note on modelling vs modeling: I assume JOSS prefers American English.
  If this is not the case, then you should keep your original `modelling`
  rather than my suggested `modeling`.

  The same applies to `Acknowledgement` (British English) vs
  `Acknowledgment` (American English).

  Feel free to reject these changes or consult with the JOSS editors.
