### 2019-09-26 76fef69
JOSS-Paper: Trim trailing whitespace
