### 2019-09-26 7d9461e
Adding citation + caption
