### 2019-09-26 a6dfd43
Paper figure referenced in text
