### 2019-09-26 5856d3e
Caption + description added for Fig 2 in paper.md
