### 2019-09-26 b828ed8
Describe cell transitions in paper.md
