+++ paper/paper.md
@@ -77,7 +77,10 @@ The simulation is adapted from a cellular automaton sand dune model, ReSCAL [@Na
 We have added features to simulate processes unique to snow, including snowfall and time-dependent cohesion.
 Fluid processes are modeled with a lattice gas cellular automaton, a method chosen to provide a good approximation to the Navier-Stokes equation at reasonable computational cost.
 Finally, the backbone of the simulation is a cellular automaton, an algorithm known to be particularly good for modeling self-organization.
-This cellular automaton contains a grid of cells of discrete types (Fig. 2, top), which evolve in nearest-neighbor pairs (doublets) as shown in Fig. 2. Each doublet transition occurs stochastically, as a poisson process with the average rates Lambda_X specified in Fig. 2.
+This cellular automaton contains a grid of cells of discrete types (Fig. 2, top), which evolve in nearest-neighbor pairs (doublets) as shown in Fig. 2. Each doublet transition occurs stochastically, as a poisson process with the average rates Lambda_X specified in Fig. 2. 
+The transitions represent the local interactions between grains that, en masse, imitate aeolian grain transport [@Narteau2014], and snow processes like metamorphosis.
+The transition rates represent the time scales of transport and metamorphism and are input parameters to the model.
+The transition grain types and orientations define the simulation 'rules', and are specified and documented in `models.c`.
 
 ![](../docs/example_images/rescal-snow_transitions.png)
 _Figure 2. Cellular automaton transitions in rescal-snow._
