### 2019-09-26 kellykochanski/rescal-snow#35 @zbeekman
JOSS Review (@zbeekman): Add CONTRIBUTING.md
The linked LAVA contributing guide is quite long. In addition, GitHub reminds users to check CONTRIBUTING.md, if it's present, when filing new issues or opening PRs.

It doesn't have to be long or complicated, but I think it would be useful to describe:

- The workflow, e.g., https://guides.github.com/introduction/flow/ vs https://datasift.github.io/gitflow/IntroducingGitFlow.html
- Expectations of contributors
- Acceptable contributions & issues
- Basic info (can just be links) on forking, commiting and opening a PR
