### 2019-09-26 kellykochanski/rescal-snow#35 @zbeekman
See https://help.github.com/articles/setting-guidelines-for-repository-contributors/ for some ideas
