### 2019-09-26 kellykochanski/rescal-snow#37 @zbeekman
JOSS-Review (@zbeekman): More prominent ack of original ReSCAL
I think you should add a more prominent acknowledgment of the original ReSCALE and its authors in the `README.md`. As it stands there is not much obvious indication of the history, heritage and relationship of Rescale-snow to ReSCALE. Furthermore, the changes relative to the original are not obvious/easy to find. I believe the GPL requires you to enumerate the changes made from the original software.

To address this I think you should add the following to the README.md:

- A more prominent description of the relationship with ReSCALE
  - Is this a fork of/from ReSCALE?
  - Which version (tag or commit SHA) of ReSCALE did you start from?
  - Are you including upstream changes from ReSCALE (if it is actively maintained) and merging these into Rescale-snow?
  - Have you considered contributing your changes back upstream?
    - A justification of the resulting decision
- A more prominent link to the Changelog/changes from the original ReSCALE

It looks like you have a change log under `docs/NEWS.md`. I would propose that you make this more legible by using bullets for sub-items rather than a one paragraph sentence per-sub item. I can submit a PR showing what I mean if you like.

---

__Optional:__

Many projects put a `CHANGELOG.md` in the root of the repository. (Or `NEWS.md`) This makes it more prominent. I personally prefer CHANGELOG.md but if the original ReSCALE uses `NEWS.md` or `NEWS` then you should probably follow that convention.

In addition, I think it would be good to move `AUTHORS.md` to the project root as well to make it more prominent.

Neither of these files are user or developer documentation in my opinion, and convention is typically locate them in the project root directly vs `./docs`.
