### 2019-09-26 ff39a7c
README: Fix ToC and section numbering
 - ToC links were broken
