+++ README.md
@@ -1,28 +1,28 @@
-# Rescal-snow <a name="introduction"></a>
+# Rescal-snow
 A model of dunes and snow waves
 
 [![status](https://joss.theoj.org/papers/822d9380fa5ce6f89cbbdebf1605df14/status.svg)](https://joss.theoj.org/papers/822d9380fa5ce6f89cbbdebf1605df14)
 
 ![](docs/example_images/snowfall_example.gif)
 
-1. [Motivation](#introduction)
-2. [Getting started](#starting)
-    1. [Prerequisites](#Prerequisites)
-    2. [Dependencies](#Dependencies)
-    3. [Download](#Download)
-    4. [Installation](#Installation)
-3. [Features](#examples)
-    1. [Example 1: a snow cone](#test-cone)
-    2. [Example 2: sintering snow](#test-sinter)
-    3. [Example 3: dune growth by snowfall](#test-snowfall)
-    4. [Example 4: parameter space exploration](#test-parallel)
-4. [References and further reading](#references)
-5. [Community guidelines](#community)
-    1. [Citation](#Citation)
-    2. [Support](#Support)
-    3. [Contributing](#contributing)
-6. [Contributors](#authors)
-7. [License](#License)
+1. [Motivation](#1-motivation)
+2. [Getting started](#2-getting-started)
+    1. [Prerequisites](#21-prerequisites)
+    2. [Dependencies](#22-dependencies)
+    3. [Download](#23-download)
+    4. [Installation](#24-installation)
+3. [Features](#3-features)
+    1. [Example 1: a snow cone](#31-sand-and-snow-dunes)
+    2. [Example 2: sintering snow](#32-sintering-snow)
+    3. [Example 3: dune growth by snowfall](#33-dune-growth-by-snowfall)
+    4. [Example 4: parameter space exploration](#34-parallel-instances-and-parameter-space-exploration)
+4. [References and further reading](#4-references-and-further-reading)
+5. [Community guidelines](#5-community-guidelines)
+    1. [Citation](#51-citation)
+    2. [Support](#52-support)
+    3. [Contributing](#53-contributing)
+6. [Contributors](#6-contributors)
+7. [License](#7-license)
 
 ### 1. Motivation
 
@@ -47,7 +47,7 @@ Rescal-snow is also designed for robust, reproducible science, and contains tool
  - A python-based workflow that manages data and analysis at runtime
 
 
-## 2. Getting started <a name="starting"></a>
+## 2. Getting started
 
 ### 2.1 Prerequisites
 
@@ -102,12 +102,12 @@ cmake -Wno-dev -DCMAKE_BUILD_TYPE=Release .. #Debug can be used instead of Relea
 make -j 4                                    #Adjust to the number of cores you have for a speedy build
 ```
 
-## 3. Features <a name="examples"></a>
+## 3. Features
 
 The examples below are fully described in the tutorial: [docs/rescal-snow-tutorial.md](docs/rescal-snow-tutorial.md).
 The [docs](docs) folder also contains descriptions of additional configuration and analysis options; check these if you're looking for functionality not found in the tutorial.
 
-### 3.1 Sand and snow dunes <a name="test-cone"></a>
+### 3.1 Sand and snow dunes
 
 The default configuration for Rescal-snow simulates snow (or sand) dune formation. This simulates processes including air flow; grain entrainment, saltation, suspension and deposition; and granular avalanches.
 
@@ -119,18 +119,18 @@ In these conditions, a pile of sand/snow (left) evolves into a dart-shaped barch
 
[…]
