### 2019-09-26 7aa0a79
README: Update CMake required to match reality
 Also add link to CMake for users who need to grab a pre-buil binary
