### 2019-09-26 9673df6
DOCS: Fix broken link & header format (inputs.md)
