+++ docs/rescal-snow-inputs.md
@@ -1,5 +1,5 @@
 
-#Inputs to rescal-snow
+# Inputs to rescal-snow
 
 Inputs controlling the behavior of rescal-snow simulations
 
@@ -25,7 +25,7 @@ The run script (run.run, snow_cone.run, etc) is a bash script. You can modify it
  - `export OMP_NUM_THREADS=1` : sets environment variable OMP_NUM_THREADS, we recommend using 1
  - `GENESIS_LOG_FILE` and `RESCAL_LOG_FILE` : set locations of two log files
  - `nice` : rescal-snow will run more slowly to avoid interfering with other applications on your computer, useful for local runs
- - run flags : 
+ - run flags :
 
 ### The parameter script
 
@@ -36,7 +36,7 @@ This file contains a list of parameters controlling simulation physics.
  - *Phys_prop_file* : points to the real data file described in the section below.
  - Simulation size *H*, *L*, *D* : sets size of simulation in cells.
  - Boundary conditions : OPEN, PERIODIC, or REINJECTION. Determines what happens to cells that pass the downwind boundary of the simulation.
- - *Lambda_X*, *Coef_X* : control the rates of cell transitions, which define/control the model physics. These are best shown graphically: see Fig. 2 of [Narteau et al., 2009](dx.doi.org/10.1029/2008JF001127).
+ - *Lambda_X*, *Coef_X* : control the rates of cell transitions, which define/control the model physics. These are best shown graphically: see Fig. 2 of [Narteau et al., 2009](https://dx.doi.org/10.1029/2008JF001127).
 
 The parameters you are likely to want to modify for snow studies include:
  - *Csp_template* : the initial conditions, see src/genesis.c
