### 2019-09-26 8709107
Ib/doc fixes (#38)
* DOCS: Fix broken links

* README: Fix link

* README: Fix ToC and section numbering

 - ToC links were broken

* DOCS: Fix broken link(s) & header

 - Also use [...]: https://... style links to make it easy to keep
   URLs in sync

* README: Update CMake required to match reality

 Also add link to CMake for users who need to grab a pre-buil binary

* DOCS: Fix link & improve legibility of s/u/readme

 - Fix a broken link to `param_space_exploration_example.py`
 - Add inline code style to executables, class names, and other code
 - Use bulleted list to increase legibility

[…]
