+++ docs/calibration_and_validation.md
@@ -1,4 +1,4 @@
-#Calibrating and validating rescal-snow
+# Calibrating and validating rescal-snow
 
 How to compare rescal-snow results with real dunes and snow features
 
@@ -17,16 +17,18 @@ rescal-snow has three non-dimensional scales:
 
 ## Calibration approach
 
-[Narteau et al, 2009](dx.doi.org/10.1029/2008JF001127) claim that each of these lengths can be calculated by matching an emergent property of the cellular automaton to an emergent property of real dunes:
+[Narteau et al, 2009] claim that each of these lengths can be calculated by matching an emergent property of the cellular automaton to an emergent property of real dunes:
 
  - The length scale can be calibrated by matching the maximum unstable wavelength of the simulation to the real maximum unstable wavelength as calculated from geomorphological theory
  - The stress scale can be calibrated by matching the ratio between wind speed and threshold wind speed from the simulation with the same ratio as estimated in reality
  - The time scale can be calculated by matching the simulated saturated flux of grains to the real saturated flux of grains
 
-This process is described in detail in [Narteau et al, 2009](dx.doi.org/10.1029/2008JF001127).  
+This process is described in detail in [Narteau et al, 2009].
 
 ## Calibration warnings and challenges
 
 This calibration presents two major model validation difficulties:
  - Changing the model parameters will change the simulated saturated fluxes, and require recalibration of the time scale.
  - The calibration relies on measurements of physical properties, such as the threshold wind speed, that may be near-impossible to measure in the field. This can add significant uncertainty to rescal-snow output.
+
+[Narteau et al, 2009]: https://dx.doi.org/10.1029/2008JF001127
