### 2019-09-26 kellykochanski/rescal-snow#41 @zbeekman
JOSS Review (@zbeekman): Tutorial part 4 (viz) is broken
The script `recolor.py` which is part of Tutorial Part 4 is broken in a number of ways:

1. There is not proper exception handling when searching for `ALTI*0_t0.log` files. When the script fails to find them it silently bails with zero status exit code (success).
2. The path to search for the `ALTI*0_t0.log` files is wrong; should be `.` not `..`.
3. The existence of PNG files checked into the repository, combined with 1 and 2 make it appear that everything worked as intended
4. Replacing the `ALTI*)_t0.log` files that come in that directory (after fixing the search path) with the files generated in tutorial step 3 results in the following python error:
```
python3 recolor.py
Traceback (most recent call last):
  File "recolor.py", line 17, in <module>
    data = np.loadtxt(os.path.join(dirname, filename))
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1141, in loadtxt
    for x in read_data(_loadtxt_chunksize):
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1065, in read_data
    % line_num)
ValueError: Wrong number of columns at line 161
```
5. Fixing the search path in `recolor.py` and re-running it with the `ALTI*0_t0.log` files results in an error after processing the first two files:
```
Recolored ALTI00050_t0.log
Recolored ALTI00150_t0.log
Traceback (most recent call last):
  File "recolor.py", line 17, in <module>
    data = np.loadtxt(os.path.join(dirname, filename))
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1141, in loadtxt
    for x in read_data(_loadtxt_chunksize):
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1065, in read_data
    % line_num)
ValueError: Wrong number of columns at line 161
```

Furthermore I recommend setting up continuous integration testing to ensure that:

- The code always builds
- Each example in the tutorial always runs
- Any additional memory leak & debugging checks can also be run provided they are not too expensive

Short of implementing, CI you should have some sort of script to ensure that all examples in the tutorial run (if you're considering that your test suite). Once you have this, it's not that hard to setup some basic CI testing.
