### 2019-09-26 68198d1
Removed broken link -> avx instructions. Fixes #40
