+++ CMakeLists.txt
@@ -76,11 +76,11 @@ function(setup_target target)
   add_sanitizers(${target})
 endfunction(setup_target)
 
-target_link_libraries(rescal    PRIVATE png pthread m)
-target_link_libraries(genesis   PRIVATE png m)  
-target_link_libraries(regenesis PRIVATE m)  
-target_link_libraries(bin2png   PRIVATE png m)  
-target_link_libraries(csp2png   PRIVATE png m)  
+target_link_libraries(rescal    PRIVATE PNG::PNG pthread m)
+target_link_libraries(genesis   PRIVATE PNG::PNG m)
+target_link_libraries(regenesis PRIVATE m)
+target_link_libraries(bin2png   PRIVATE PNG::PNG m)
+target_link_libraries(csp2png   PRIVATE PNG::PNG m)
 
 setup_target(rescal)
 setup_target(genesis)
