### 2019-09-26 3443ba6
Bullet points in NEWS.md #37
Added bullet points for clearer formatting, specifically mentioned split from ReSCAL 1.6.
