+++ docs/NEWS.md
@@ -1,136 +1,139 @@
 # News and development history
 
 ## rescal-snow 1.0
+
+Split from ReSCAL. Code slimmed down to focus on DUN and SNO models, rather than RIV, CMB, etc.
+
 #### Models, templates and functionality
 
-Added SNO model
+- Added SNO model
 
-Added snowfall ability to SNO model, with accompanying Csp_template SNOWFALL
+- Added snowfall ability to SNO model, with accompanying Csp_template SNOWFALL
 
-Added sintering ability to SNOW model
+- Added sintering ability to SNOW model
 
-Removed non-dune-related models including RIV, CMB, etc
+- Removed non-dune-related models including RIV, CMB, etc
 
 #### Tools for set-up, post-processing, and parameter space exploration
 
-Added 'scripts/utilities' folder containing python tools for setting up large numbers of runs for parameter space explorations
+- Added 'scripts/utilities' folder containing python tools for setting up large numbers of runs for parameter space explorations
 
-Added cross-correlation analysis tool
+- Added cross-correlation analysis tool
 
-Added fft analysis and visualization tool
+- Added fft analysis and visualization tool
 
 #### Organization
 
-Removed many regions of commented-out code
+- Removed many regions of commented-out code
 
-Model output now funnelled through 'write_output' files for standardization
+- Model output now funnelled through 'write_output' files for standardization
 
-Model output may now be organized in output directories
+- Model output may now be organized in output directories
 
 #### Performance and compatibility
 
-Converted all instances of 'int' to 'int32_t' for cross-platform compatibility
+- Converted all instances of 'int' to 'int32_t' for cross-platform compatibility
 
-Added additional AVX optimizations to the lattice gas
+- Added additional AVX optimizations to the lattice gas
 
-Removed GTK dependency; graphics now handled entirely through LIBPNG
+- Removed GTK dependency; graphics now handled entirely through LIBPNG
 
-Modified configure and make pathways to account for new dependencies
+- Modified configure and make pathways to account for new dependencies
 
-## ReSCAL Version 1.6
+## ReSCAL Version 1.6 (2016)
 
-AVX optimization of lattice-gas propagation.
+- AVX optimization of lattice-gas propagation.
 
-Optimization of flow interpolation.
+- Optimization of flow interpolation.
 
-New command-line option (-pat) for the concurrent execution of stochastic engine 
+- New command-line option (-pat) for the concurrent execution of stochastic engine 
 and lattice gas, with non-deterministic algorithm.
 
-Command-line help improved with new options -h and -hm.
+- Command-line help improved with new options -h and -hm.
 
-More CSP templates and parameters for DUN model.
+- More CSP templates and parameters for DUN model.
 
-Various fixes.
+- Various fixes.
 
 ## ReSCAL Version 1.5
-Relief light-shading with various colors, for colored sand and topographic 
+- Relief light-shading with various colors, for colored sand and topographic 
 obstacles (DUM cells) in DUN and AVA models, and for fluid cells in RIV model.
 
-More simple syntax for graphical options in command-line.
+- More simple syntax for graphical options in command-line.
 
-The graphical display readapts automatically as the window gets larger.
+- The graphical display readapts automatically as the window gets larger.
 
-Better interactivy with the positioning of cross sections.
+- Better interactivity with the positioning of cross sections.
 
-New TRACE_FLUX mode in DUN model for the calculation of the average sand flux
+- New TRACE_FLUX mode in DUN model for the calculation of the average sand flux
 over time, compatible with rotations.
 
-New option for stability analysis in DUN model.
+- New option for stability analysis in DUN model.
 
[…]
