### 2019-09-26 12d3baf
Add statement about ReSCAL to README #37
