### 2019-09-26 kellykochanski/rescal-snow#37 @zbeekman
I think I misspelled ReSCAL everywhere, but my meaning should be clear enough.
