### 2019-09-26 kellykochanski/rescal-snow#37 @kellykochanski
(Links to NEWS currently broken; it'll work once NEWS is moved.)
