### 2019-09-26 #1699 @zbeekman
@kellykochanski I just wanted to update you on my progress. (I know you're busy responding to issues & coding so no need to respond.) So far I have:

- Built & tested on macOS w/ native & Homebrew tooling
  - Opened https://github.com/kellykochanski/rescal-snow/pull/42 to resolve building with brewed gccs
- Thoroughly reviewed documentation and submitted some PRs for broken links etc.
  - Very nice documentation documents & tutorial
- Thoroughly reviewed the paper
  - Again, very nice!

To conclude the review I will do the following (out of time today):

- Build and test on Linux with Intel compilers & gcc
- Perform the parallel parameter space study in the tutorial
- Verify (quick and dirty) performance numbers in the parallelism/performance document

Outstanding issues (so far) that I think need to be resolved, in order of severity:

- [ ] https://github.com/kellykochanski/rescal-snow/issues/41
- [ ] https://github.com/kellykochanski/rescal-snow/issues/35
- [ ] https://github.com/kellykochanski/rescal-snow/issues/37

Some CI testing and/or single test script to verify the tutorial works would have probably prevented https://github.com/kellykochanski/rescal-snow/issues/41 and would be really good to add. (You probably cannot do the last parametric parameter study example on a free CI solution, but a script for including this could be a good supplement to CI testing here.

So far things look really good! I think https://github.com/kellykochanski/rescal-snow/issues/41 definitely should be addressed/resolved before publication, but most of the other things are pretty minor.
