### 2019-09-26 kellykochanski/rescal-snow#41 @kellykochanski
@defaziogiancarlo Quick question - do you have a preferred function for visualizing an ALTI file?

I saw that your last commit, after I wrote the now-broken piece of the tutorial, included some new tools for reading and visualizing Rescal height maps / ALTI files. (At present, there are several ways to do this - it would be nice to condense onto one working function if possible)
