### 2019-09-26 fed9eed
Moved NEWS to root file #37
