### 2019-09-26 kellykochanski/rescal-snow#37 @kellykochanski
AUTHORS.md is also in the root directory.
