### 2019-09-26 kellykochanski/rescal-snow#37 @kellykochanski
@zbeekman I believe the points raised in this issue are now addressed.
