### 2019-09-26 kellykochanski/rescal-snow#37 @zbeekman
Looks good to me. I'll close this. I commented on the `NEWS.md` file: I think removing the extra space between successive bullets would improve legibility, but at this point that's just a nit; adjust the style however you see fit.
