+++ CMakeLists.txt
@@ -8,8 +8,9 @@ project (rescal-snow
 set(CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/cmake" ${CMAKE_MODULE_PATH})
 find_package(Sanitizers)
 
-find_package(ZLIB 1.2 REQUIRED)
+set(THREADS_PREFER_PTHREAD_FLAG TRUE)
 find_package(PNG 1.6 REQUIRED)
+find_package(Threads )
 
 set(COMMON_SOURCES
   src/cells.c
@@ -76,7 +77,7 @@ function(setup_target target)
   add_sanitizers(${target})
 endfunction(setup_target)
 
-target_link_libraries(rescal    PRIVATE PNG::PNG pthread m)
+target_link_libraries(rescal    PRIVATE PNG::PNG Threads::Threads m)
 target_link_libraries(genesis   PRIVATE PNG::PNG m)
 target_link_libraries(regenesis PRIVATE m)
 target_link_libraries(bin2png   PRIVATE PNG::PNG m)
