### 2019-09-27 0d45cca
Drop zlib requirement from README since it is implied by PNG
