### 2019-09-27 e1212f3
Only attempt linking pthreads if present
