+++ CMakeLists.txt
@@ -77,7 +77,11 @@ function(setup_target target)
   add_sanitizers(${target})
 endfunction(setup_target)
 
-target_link_libraries(rescal    PRIVATE PNG::PNG Threads::Threads m)
+if(TARGET Threads::Threads)
+  target_link_libraries(rescal    PRIVATE PNG::PNG Threads::Threads m)
+else()
+  target_link_libraries(rescal    PRIVATE PNG::PNG m)
+endif()
 target_link_libraries(genesis   PRIVATE PNG::PNG m)
 target_link_libraries(regenesis PRIVATE m)
 target_link_libraries(bin2png   PRIVATE PNG::PNG m)
