+++ CMakeLists.txt
@@ -10,7 +10,7 @@ find_package(Sanitizers)
 
 set(THREADS_PREFER_PTHREAD_FLAG TRUE)
 find_package(PNG 1.6 REQUIRED)
-find_package(Threads )
+find_package(Threads REQUIRED)
 
 set(COMMON_SOURCES
   src/cells.c
@@ -77,11 +77,7 @@ function(setup_target target)
   add_sanitizers(${target})
 endfunction(setup_target)
 
-if(TARGET Threads::Threads)
-  target_link_libraries(rescal    PRIVATE PNG::PNG Threads::Threads m)
-else()
-  target_link_libraries(rescal    PRIVATE PNG::PNG m)
-endif()
+target_link_libraries(rescal    PRIVATE PNG::PNG Threads::Threads m)
 target_link_libraries(genesis   PRIVATE PNG::PNG m)
 target_link_libraries(regenesis PRIVATE m)
 target_link_libraries(bin2png   PRIVATE PNG::PNG m)
