### 2019-09-27 6bad9e1
Create CONTRIBUTING.md
