### 2019-09-27 a3889cc
Formatting in CONTRIBUTING.md
