+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-#How to contribute <a name="contribute">
+# How to contribute <a name="contribute">
 
 Thank you for taking the time to contribute! We're excited to see your work.
 
@@ -8,15 +8,13 @@ Thank you for taking the time to contribute! We're excited to see your work.
     - [Git workflow and pull requests](#workflow)
  [Code of conduct](#conduct)
  
- <a name="bugs">
- ##Reporting bugs
+ ##Reporting bugs <a name="bugs">
  Please report bugs as Github Issues. Include steps necessary to reproduce the bug, any error text, and a description of the system on which the bug occurred. 
  For installation issues, include version numbers for the compiler, cmake, and the library dependencies.
  
  If you think you can fix the bug, great! See the features and pull request guidelines below for how to incorporate your bugfix into the Rescal-snow github.
  
- <a name="features">
- ##Adding features
+ ## Adding features <a name="features">
  Rescal-snow grows by adding new features. Features include:
   - New capabilities for the Rescal-snow simulation (e.g. cohesion, or new inputs)
   - New workflow management, analysis, or visualization scripts
@@ -35,9 +33,8 @@ To make this possible, features that add new capabilities to Rescal-snow should 
   - A tutorial or example in `docs` explaining the use of any novel capabilities
   - A reference that makes new capabilities discoverable through either `README.md` or `docs/rescal-snow-tutorial.md`
   - For particularly significant changes (project >1wk), appropriate additions to `NEWS.md` and `AUTHORS.md`
-  
- <a name="workflow">
- ###Git workflow and pull requests
+ 
+ ### Git workflow and pull requests <a name="workflow">
  We use the git branching workflow described [here](https://nvie.com/posts/a-successful-git-branching-model/) (if new to git, see link for useful commands).
  
  This consists of 4 main types of branches:
@@ -55,15 +52,15 @@ and commit your changes.
 
 When the feature or bugfix is ready to be incorporated into Rescal-snow, open a [pull request](https://help.github.com/en/articles/creating-a-pull-request) against `develop`.
  
- <a name="conduct">
- #Contributor Covenant Code of Conduct 
-##Our Pledge
+ # Contributor Covenant Code of Conduct <a name="conduct">
+ 
+## Our Pledge
 In the interest of fostering an open and welcoming environment, we as contributors and maintainers pledge to make
 participation in our project and our community a harassment-free experience for everyone, regardless of age, body size,
 disability, ethnicity, sex characteristics, gender identity and expression, level of experience, education, socio-economic status,
 nationality, personal appearance, race, religion, or sexual identity and orientation. 
 
-##Our Standards 
+## Our Standards 
 Examples of behavior that contributes to creating a positive environment include: 
 - Using welcoming and inclusive language
 - Being respectful of differing viewpoints and experiences
@@ -78,20 +75,20 @@ Examples of unacceptable behavior by participants include:
 - Publishing others' private information, such as a physical or electronic address, without explicit permission
 - Other conduct which could reasonably be considered inappropriate in a professional setting 
 
-##Our Responsibilities 
+## Our Responsibilities 
 - Project maintainers are responsible for clarifying the standards of acceptable behavior and are expected to take appropriate
 and fair corrective action in response to any instances of unacceptable behavior. 
 - Project maintainers have the right and responsibility to remove, edit, or reject comments, commits, code, wiki edits, issues,
 and other contributions that are not aligned to this Code of Conduct, or to ban temporarily or permanently any contributor for
 other behaviors that they deem inappropriate, threatening, offensive, or harmful. 
 
-##Scope 
+## Scope 
 This Code of Conduct applies within all project spaces, and it also applies when an individual is representing the project or its
[…]
