### 2019-09-27 kellykochanski/rescal-snow#35 @kellykochanski
@zbeekman Added contributing.md:
https://github.com/kellykochanski/rescal-snow/blob/kk/JOSS-fixes/CONTRIBUTING.md
