### 2019-09-27 b3ed836
Updated 'contributing' section of README #35
