### 2019-09-27 kellykochanski/rescal-snow#35 @zbeekman
@kellykochanski Awesome!

Two things:

1. If you are adding a code of conduct (contributor covenant, etc.) it should go in its own CODE_OF_CONDUCT.md file in the root of repository. This is common practice and GitHub will find it and understand that its there & enable bells and whistles associated with it. While I did not ask for a CODE_OF_CONDUCT.md, I do think it's always good to establish expectations for treating each other fairly & professionally so I appreciate that you added one!
2. It looks like you need a space in your H2 heading between `##` and the text. Older github flavored markdown (GFM) used to allow this without a space but the new rendering engine does not.

Looking good. Once you fix the formatting and separate the `CODE_OF_CONDUCT.md` (Or delete it if you really want to although I would discourage that...) I think this issue is resolved.
