### 2019-09-27 1fe3c1b
Updated visualization exercise in tutorial #41
- Exercise now active (users do more steps in python, useful for gaining familiarity with directory structure and utilities)
- Uses functions from heightmap.py instead of one-off recolor.py script
