### 2019-09-27 69fa5d1
Updated example images for visualization tutorial #41
