### 2019-09-27 8b8c621
Separated CONTRIBUTING and CODE_OF_CONDUCT #35
