### 2019-09-27 kellykochanski/rescal-snow#35 @kellykochanski
@zbeekman Great! I've separated the files and fixed the typo. Marking this closed.
