### 2019-09-27 kellykochanski/rescal-snow#41 @kellykochanski
@zbeekman The tutorial will now work as shown; it now uses a mix of explicit python commands (to walk users through some of the output files + tools) and a re-usable tool from `heightmap.py` in the utilities folder instead of using `recolor.py`, which was a hacky one-off script just for this tutorial.
