### 2019-09-27 kellykochanski/rescal-snow#41 @defaziogiancarlo
The `scripts/utilities/heightmap.py` script has a few ways to draw height maps.
All the ways are just using matplotlib default settings.
If you create a `HeightMap` object, using `heightmap.HeightMap("Some ALTI file")`
There's a `draw` method and a `plot_3d`, both of which make a window pop up.

The `CellSpace` objects contain `HeightMap` objects, you can use the `draw_height_map` method.
