### 2019-09-27 kellykochanski/rescal-snow#35 @zbeekman
LGTM!
