### 2019-09-27 kellykochanski/rescal-snow#41 @kellykochanski
@zbeekman I've added a [test script](test/test.sh) which runs 3 of the examples. It is not fully automatic - after it runs things, it repeatedly prompts the user to confirm that output looks OK - but since so much of the output of Rescal-snow I believe that keeping human eyes on the output will be useful, and this does make it much easier to check that large swaths of the code still work after changes.

Is this a reasonable level of testing? I did some googling for CI for github, but it wasn't encouraging.

(I included only one of the Rescal-snow examples, the snow cone, because in my experience all 3 crash or succeed together, and including only one makes the test faster to run and thus easier to use).
