### 2019-09-27 1ad1720
Documenting test script
