### 2019-09-27 kellykochanski/rescal-snow#41 @r-barnes
@kellykochanski : I can help with this, but it would be good if your could merge #42 first.
