### 2019-09-27 f815023
CMake: Fix macOS GCC compilation: Link PNG::PNG (#42)
* CMake: Fix macOS GCC compilation: Link PNG::PNG

  - On macOS builds with brewed gcc8 and gcc9 were failing to link libpng
    - I guess Apple's clang/cc have this stored away in Xcode paths somewhere
  - To find my brewed libpng correctly one cannot simply pass `png` to
    `target_link_libraries()`
  - Since minimum allowed CMake is 3.9, take advantage of transitive linked
    target property propagation and link to the exported target, `PNG::PNG`
    defined by the intrinsic `FindPNG.cmake` CMake module

* Drop ZLIB prefer Threads::Thr
[…]
