### 2019-09-27 kellykochanski/rescal-snow#41 @kellykochanski
Merged.
