### 2019-09-27 5cbeadf
Fix up testing
