### 2019-09-27 kellykochanski/rescal-snow#41 @r-barnes
See #43.
