### 2019-09-27 1768c6b
Restored semi-manual test for utilities
