### 2019-09-27 aea6414
Updated docstrings in test to match content #43
