### 2019-09-27 9877de9
Drop python3 param_space_exploration_example.py from test-utilities.sh
