### 2019-09-29 a6c6a95
Fix up testing (#43)
* Fix up testing

* Fixing stochastic CA seed in test #43

* Restored semi-manual test for utilities

* Updated docstrings in test to match content #43

* Drop python3 param_space_exploration_example.py from test-utilities.sh
