+++ test/test-utilities.sh
@@ -12,37 +12,20 @@
 # Success at each stage is confirmed by the user.
 
 
-echo "Attempting to build Rescal-snow..."
+echo "Building Rescal-snow..."
 cd ..
 mkdir build
 cd build
 cmake -Wno-dev -DCMAKE_BUILD_TYPE=Release ..
 make -j 16
 cd ..
-while true; do
-	read -p "Did Rescal-snow build successfully? [y/n]" yn
-	case $yn in
-		[Yy]* ) echo "Good! Continuing."; break;;
-		[Nn]* ) exit;;
-		* ) echo "Please input [y/n] to continue.";;
-	esac
-done
 
-# Test that Rescal-snow runs
+# Run snow-cone test to generate example data
 echo "Attempting to run Rescal-snow using the snow_cone test case..."
 echo "(This may take a few minutes)"
+echo "(Output not displayed - use test-rescal.sh if necessary)"
 cd scripts
 ./snow_cone.run
-eog *.png
-while true; do
-	read -p "Does the output look good? [y/n]" yn
-	case $yn in
-		[Yy]* ) echo "Good! Continuing."; break;;
-		[Nn]* ) exit;;
-		* ) echo "Please input [y/n] to continue.";;
-	esac
-done
-
 
 # Test the visualization example
 echo "Attempting visualization example..."
@@ -52,25 +35,6 @@ rm viz_example_test.py
 eog out/*image.png
 while true; do
 	read -p "Does the output look good? [y/n]" yn
-	case $yn in
-		[Yy]* ) echo "Good! Continuing."; break;;
-		[Nn]* ) exit;;
-		* ) echo "Please input [y/n] to continue.";;
-	esac
-done
-
-# Test rescal_utilities for setting up parallel runs
-echo "Attempting rescal_utilities example..."
-cd utilities
-python3 param_space_exploration_example.py
-cd ../..
-echo "Printing parallel directories:"
-ls test_parallel_runs
-echo "Printing contents of one directory:"
-ls test_parallel_runs/tauMin0_lambdaI0.001
-rm -rf test_parallel_runs
-while true; do
-	read -p "Did Rescal-snow set up 10 parallel directories, each with rescal, genesis, real_data, run.run and run.par? [y/n]" yn
 	case $yn in
 		[Yy]* ) echo "Good! Continuing."; break;;
 		[Nn]* ) exit;;
