### 2019-09-30 kellykochanski/rescal-snow#41 @kellykochanski
@zbeekman Rescal-snow now has a [test](https://github.com/kellykochanski/rescal-snow/tree/kk/JOSS-fixes/test) directory, containing scripts that test:

- [The build system, utilities and an example simulation](https://github.com/kellykochanski/rescal-snow/blob/kk/JOSS-fixes/test/test-rescal.sh)
- [The visualization example that triggered this issue](https://github.com/kellykochanski/rescal-snow/blob/kk/JOSS-fixes/test/test-utilities.sh)

We haven't set up CI (partly due to time, partly because the visualization example isn't quite automated and requires user input), but these changes have fixed the bug you found, and created a more organized testing process to lower the risk of another one.
