### 2019-10-03 #1699 @kellykochanski
@zbeekman I believe these issues have now been addressed; issues 35 and 37 are closed, and it would be great if you could confirm our work on [41](https://github.com/kellykochanski/rescal-snow/issues/41).

I'm curious to see your results for the performance test.
