### 2019-10-14 #1699 @zbeekman
On the agenda for today! Thanks for the nudge.

On Mon, Oct 14, 2019 at 10:56 AM Kelly Kochanski <notifications@github.com>
wrote:

> @zbeekman <https://github.com/zbeekman> Do you know when you will have
> time to look at this?
>
