### 2019-10-14 kellykochanski/rescal-snow#41 @zbeekman
@kellykochanski I'm still hitting an error in `heightmap.py` when following the updated tutorial in the (now-manual) python section:

```
Traceback (most recent call last):
  File "<stdin>", line 4, in <module>
  File "utilities/heightmap.py", line 242, in __init__
    self.height_map = np.loadtxt(height_input).astype(np.uint8)
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1146, in loadtxt
    for x in read_data(_loadtxt_chunksize):
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1071, in read_data
    % line_num)
ValueError: Wrong number of columns at line 161
```

Here are my complete commands.

```
$ python3
Python 3.7.4 (default, Sep  7 2019, 18:27:02)
[Clang 10.0.1 (clang-1001.0.46.4)] on darwin
>>> import sys
>>> sys.path.append('utilities')    # Add utilities to Python path
>>> import os
>>> import matplotlib.pyplot as plt # Plotting
>>> import heightmap                # utilities/heightmap.py contains tools for reading and visualizing height maps (ALTI files)
>>> # Find the heightmap files to visualize. In this tutorial, they are named out/ALTIxxxxx.log.
... filenames = ['out/'+f for f in os.listdir('out') if f[:4]=='ALTI']
>>> filenames = [f for f in filenames if f[-4:]=='.log']
>>> filenames.sort()
>>> print(filenames)
['out/ALTI00000_t0.log', 'out/ALTI00001_t0.log', 'out/ALTI00002_t0.log', 'out/ALTI00003_t0.log', 'out/ALTI00004_t0.log', 'out/ALTI00005_t0.log', 'out/ALTI00006_t0.log', 'out/ALTI00007_t0.log', 'out/ALTI00008_t0.log', 'out/ALTI00009_t0.log', 'out/ALTI00010_t0.log', 'out/ALTI00011_t0.log', 'out/ALTI00012_t0.log', 'out/ALTI00013_t0.log', 'out/ALTI00014_t0.log', 'out/ALTI00015_t0.log', 'out/ALTI00016_t0.log', 'out/ALTI00017_t0.log', 'out/ALTI00018_t0.log', 'out/ALTI00019_t0.log', 'out/ALTI00020_t0.log', 'out/ALTI00021_t0.log', 'out/ALTI00022_t0.log', 'out/ALTI00023_t0.log', 'out/ALTI00024_t0.log', 'out/ALTI00025_t0.log', 'out/ALTI00026_t0.log', 'out/ALTI00027_t0.log', 'out/ALTI00028_t0.log', 'out/ALTI00029_t0.log', 'out/ALTI00030_t0.log', 'out/ALTI00031_t0.log', 'out/ALTI00032_t0.log', 'out/ALTI00033_t0.log', 'out/ALTI00034_t0.log', 'out/ALTI00035_t0.log', 'out/ALTI00036_t0.log', 'out/ALTI00037_t0.log', 'out/ALTI00038_t0.log', 'out/ALTI00039_t0.log', 'out/ALTI00040_t0.log', 'out/ALTI00041_t0.log', 'out/ALTI00042_t0.log', 'out/ALTI00043_t0.log', 'out/ALTI00044_t0.log', 'out/ALTI00045_t0.log', 'out/ALTI00046_t0.log', 'out/ALTI00047_t0.log', 'out/ALTI00048_t0.log', 'out/ALTI00049_t0.log', 'out/ALTI00050_t0.log', 'out/ALTI00051_t0.log', 'out/ALTI00052_t0.log', 'out/ALTI00053_t0.log', 'out/ALTI00054_t0.log', 'out/ALTI00055_t0.log', 'out/ALTI00056_t0.log', 'out/ALTI00057_t0.log', 'out/ALTI00058_t0.log', 'out/ALTI00059_t0.log', 'out/ALTI00060_t0.log']
>>> # Visualize the files. This may take a few seconds.
... for name in filenames:
...     # Read the file as a HeightMap object
...     hm  = heightmap.HeightMap(name)
...     # hm.draw                   # Uncomment to view each image as it is produced
...     hm.save_color_map(name[:-4] + '_image.png')
...
Traceback (most recent call last):
  File "<stdin>", line 4, in <module>
  File "utilities/heightmap.py", line 242, in __init__
    self.height_map = np.loadtxt(height_input).astype(np.uint8)
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1146, in loadtxt
    for x in read_data(_loadtxt_chunksize):
  File "/usr/local/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1071, in read_data
    % line_num)
ValueError: Wrong number of columns at line 161
```

I can try this on a Linux box to see if this is some strange macOS behavior. I still think that if you have this script as part of your tutorial, there should be a way to test it. If you want it to be more interactive than just running a script I recommend exploring jupyter notebooks, and their ability to be run non-interactively for testing.

If I can't reproduce this issue on linux, then I may just consider this good enough, and a macOS specific bug (from the perspective of my JOSS reviewer duties) but in the long run it would be good to see testing on other platforms of this script.

Also, I'll comment in the PR, but I think you should re-word the python section of the tutorial a little bit: you're no longer including a script, but rather giving folks a set of commands to run in an interactive interpreter.
