### 2019-10-14 kellykochanski/rescal-snow#41 @zbeekman
I wonder if it might make sense to use [pipenv](https://github.com/pypa/pipenv) or conda's `environment.yml` to specify python versions and lock required module versions? This would increase re-producibility and headaches caused by wacky python environments.
