### 2019-10-14 kellykochanski/rescal-snow#44 @zbeekman
JOSS Review (@zbeekman): Reproducibility of Python Environments
EDITED: Wrong issue linked

For reproducibility, it would be good to specify the python requirements and perform testing in a reproducible environment using [`pipenv`] or conda, or at the very least using a `requirements.txt` file pinning versions (less preferable--[`pipenv`] seems to be the preferred method moving forwards).

Rather than keep discussing the python errors I am encountering in the tutorial as part of #41, I thought this was important enough to break out into a separate issues.

While following along with the tutorial, once I get to the Python visualization section the script/interactive commands keep dying for me. I'm on macOS (Mojave) using Python 3.7.4. It is possible/likely that the issues I'm encountering are due to differences in my Python version, or differences in the python module versions I'm using.

The error I am gettings is:
```
Traceback (most recent call last):
  File "<stdin>", line 4, in <module>
  File "utilities/heightmap.py", line 242, in __init__
    self.height_map = np.loadtxt(height_input).astype(np.uint8)
  File "/Users/ibeekman/.local/share/virtualenvs/scripts-LccjI4Uk/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1146, in loadtxt
    for x in read_data(_loadtxt_chunksize):
  File "/Users/ibeekman/.local/share/virtualenvs/scripts-LccjI4Uk/lib/python3.7/site-packages/numpy/lib/npyio.py", line 1071, in read_data
    % line_num)
```

The code I am running is:
```
import sys
sys.path.append('utilities')    # Add utilities to Python path
import os
import matplotlib.pyplot as plt # Plotting
import heightmap                # utilities/heightmap.py contains tools for reading and visualizing height maps (ALTI files)

# Find the heightmap files to visualize. In this tutorial, they are named out/ALTIxxxxx.log.
filenames = ['out/'+f for f in os.listdir('out') if f[:4]=='ALTI']
filenames = [f for f in filenames if f[-4:]=='.log']
filenames.sort()
print(filenames)

# Visualize the files. This may take a few seconds.
for name in filenames:
    # Read the file as a HeightMap object
    hm  = heightmap.HeightMap(name)
    # hm.draw                   # Uncomment to view each image as it is produced
    hm.save_color_map(name[:-4] + '_image.png')
```

From my very brief experimentation, it seems like the following python packages are required:

- matplotlib
- numpy (already required by matplotlib, but used explicitly)
- scipy

After generating a virtual environment using [`pipenv`] and activating it, I still encountered the same errors.

Note: Github only likes certain file formats, so I gzipped these to allow me to upload them. I can provide them as `.txt` if you need/want me to.

[Pipfile](https://github.com/kellykochanski/rescal-snow/files/3726348/Pipfile.gz)
[Pipfile.lock.gz](https://github.com/kellykochanski/rescal-snow/files/3726353/Pipfile.lock.gz)

[`pipenv`]: https://github.com/pypa/pipenv
