### 2019-10-14 kellykochanski/rescal-snow#41 @defaziogiancarlo
This looks like one of the `ALTI` files might be malformed. Using `np.loadtxt`, I get  a
`ValueError: Wrong number of columns at line <number>`
when I try load in a file in which I purposely removed a value from on of the lines.

Do any of the `_image.png` files get created, or does it fail on the first `ALTI` file?
