### 2019-10-14 kellykochanski/rescal-snow#41 @zbeekman
@defaziogiancarlo could this be a line ending error/issue? People often setup git to checkout text files with native endings, etc. So I wonder if there's logic somewhere that assumes a certain number of characters/bytes per line? Then if you convert the line endings from `\n\r` to `\n` (or vise versa) something blows up?

I have no png files in `./script` or in `./script/out`.
