### 2019-10-14 kellykochanski/rescal-snow#41 @zbeekman
Hmmmm, never mind, I see that the ALTI files were generated by running the code, and are *not* checked in.

Here is a tarball of my ALTI files, if it helps. I'm following along the tutorial.

[ALTI_t0-logs.tar.gz](https://github.com/kellykochanski/rescal-snow/files/3726618/ALTI_t0-logs.tar.gz)
