### 2019-10-14 kellykochanski/rescal-snow#41 @defaziogiancarlo
Something is wrong with the `ALTI` file generation. The number of columns doubles for some reason at line 161. I think the actual data ends at line 160 and the rest is all 6 and 7, which is neutral and sand input cell types.
