### 2019-10-14 kellykochanski/rescal-snow#41 @kellykochanski
@defaziogiancarlo @zbeekman This is a bug in the file management.

When you run all of the tutorials in sequence, not all of the output from old runs (snow_cone, sintering) is deleted before hitting the `snowfall` tutorial that produce the output used in the visualization tutorial.

Thus, the visualization tutorial is trying to read ALTI files from a mix of runs. This triggers an error, as `heightmap` expects to read files with a consistent shape, size, and layout. Presumably I missed this earlier by testing the tutorials out of order.

I've updated the tutorial scripts to clean the entirety of the old output before generating new output, which should remove the problem.

I'm also updating `heightmap` to produce a more useful error message in the case of irregular file shapes and sizes.
