### 2019-10-14 1f286c7
Updating clean to prevent ALTI frankenfiles
