### 2019-10-14 c7032c7
#41 better error message for heightmap read
