### 2019-10-14 01d6924
Clarifying language in ./clean tutorial instruction
