### 2019-10-15 c8c3f77
Make test scripts more robust & portable
