### 2019-10-15 6ed080b
Do not output gifs by default
 - Commented out gif creation script call, reenable if you like
