### 2019-10-15 c82bb8b
Match no-cleanup for test-utilities.sh on kk/JOSS
