### 2019-10-15 ca6f16f
Add back python script needed by test-utilities.sh
