### 2019-10-15 a5c42dd
Removed copy-paste errors from tutorial
