### 2019-10-15 ca67af7
Better formatting for parallel run errors in tutorial
