+++ docs/rescal-snow-tutorial.md
@@ -356,10 +356,17 @@ The runs with high wind speed (low Tau\_min) have less even snow cover, with bet
 
 >*Parallel run errors?*
 >
-> - msub/sbatch commands not allowed by your computing cluster -> seek support from someone familiar with the cluster
-> - permissions errors -> run `chmod u+rwx \*` in test_parallel_runs; contact administrator if this is disallowed
-> - script could not find input (run.run or run.par or sealevel_snow.prop) -> rerun param_space_exploration_example.py and confirm that it produced the output above; check that you're running msub/sbatch from scripts; check relative directory references in msub/sbatch script and submit.sh
-> - runs time out before producing useful output -> increase walltime; large simulations on slow machines may ultimately take a few hours
+> - msub/sbatch commands not allowed by your computing cluster 
+>      * seek support from someone familiar with the cluster
+> - permissions errors 
+>      * run `chmod u+rwx \*` in test_parallel_runs directory 
+>      * contact administrator if this is disallowed
+> - script could not find input (run.run or run.par or sealevel_snow.prop)
+>      * rerun param_space_exploration_example.py and confirm that it produced the output above
+>      * check that you're running msub/sbatch from scripts
+>      * check relative directory references in msub/sbatch script and submit.sh
+> - runs time out before producing useful output
+>      * increase walltime; large simulations on slow machines may ultimately take a few hours
 
 ### Run ReSCAL within python <a name="python"></a>
 
