### 2019-10-15 6ff3dd9
Update docs/rescal-snow-tutorial.md to reflect updated viz example
Co-Authored-By: Izaak "Zaak" Beekman <zbeekman@gmail.com>
