### 2019-10-15 15e132f
Updating permissions directions, script names for parallel example #45
