### 2019-10-15 kellykochanski/rescal-snow#44 @zbeekman
merging the appropriate parts of #46 should address this.
