### 2019-10-15 kellykochanski/rescal-snow#41 @zbeekman
@kellykochanski Ok, I've gotten through the entire tutorial. There may be an additional issue with the parameter space exploration example. I'm on a slurm system, and using your scripts, as outlined in the tutorial, appears to only run the first example. Furthermore, the sbatch script doesn't look like a typical slurm array job, I suspect this is the source of the issue.
