### 2019-10-16 #1699 @zbeekman
I think I have more or less concluded my review. There are many great features of rescal-snow, and, in most circumstances, building it was straight forwards. I've tested it on:

- macOS with clang
- macOS with GCC 9
- HPC cluster with GCC 6
- HPC cluster with Intel 18

In general there is a fairly large quantity of usage documentation and tutorials. At times this has been a double edged sword because the documentation and tutorials are markdown documents which can become out of sync with the project and software. Overall, the software testing and processes relating to reproducibility and quality assurance do leave some room for improvement. Despite this, these areas have improved over the course of this review, which is encouraging.

*I am recommending publication with minor revisions*, discussed in https://github.com/kellykochanski/rescal-snow/issues/41 and https://github.com/kellykochanski/rescal-snow/issues/44. The *main* software package seems fairly robust and has some really interesting capabilities. Some of the supporting scripts and pre/post-processing tools do not appear to be as robust as I would like, and caused me to devote more time than I would have liked to troubleshooting. Nevertheless, I feel that the overall package meets the requirements for JOSS publication. The weakest points, in my opinion, are in the `Documentation` section, in particular, __Automated Tests__ and __Functionality Documentation__. The __Functionality Documentation__ issues were because following the examples did not produce the results they claimed to due to bugs, most of which are fixed.

While I do not believe the aforementioned issues preclude rescal-snow from being published--i.e., "major revisions",--I recommend that the authors take additional steps (now or after publication) to reduce the brittleness and manual nature of a lot of the testing and documentation:

1. Setup CI testing
   - test on linux
   - test on macOS
   - test scripts (Bash and Python) used in tutorials and user-documentation
   - setup unit tests for Python modules/code
2. Setup static analysis (you can do this during CI) of the shell scripts and python modules & scripts as part of your quality assurance process
   - [ShellCheck](https://github.com/koalaman/shellcheck#installing) is a great tool for shell scripts
3. Consider transitioning shell scripts to either Python with robust exception handling or CMake tasks via custom targets and/or ctest tests
   - Python with exception handling is more robust and portable
   - Relying on CMake rather than *nix tools and bash scripts make it more likely to run on all major OSes and the CMake code can be exercised during CI testing or by quickly typing `ctest`.
4. Restructure some or all of the documentation, most notably the tutorial, as a Jupyter Notebook.
   - Test the notebook(s) during CI to ensure code changes haven't broken the examples and tutorials
   - This was probably the greatest source of delay and frustration during the review. Automating ways to check the code in the documentation and keeping code with documentation will help ensure things do not get out of sync.

Great job with rescal-snow!
