### 2019-10-16 98b3108
More robust test scripts, GIF maker script, robust & reproducible python environment (#46)
* Make test scripts more robust & portable

* Add script to create PNGs with imagemagick

* Do not output gifs by default

 - Commented out gif creation script call, reenable if you like

* Add pipenv environment description and lockfile

 - Ensure reproducible virtual envs

* Add requirements.txt for old-school pip/venv users

* Match no-cleanup for test-utilities.sh on kk/JOSS

* Add back python script needed by test-utilities.sh

* Handling CMake policies better & handle cmp0074

 - cmp0074 let's you 
[…]
