+++ requirements.txt
@@ -1,0 +1,10 @@
+cycler==0.10.0
+kiwisolver==1.1.0
+matplotlib==3.1.1
+numpy==1.17.2
+pandas==0.25.1
+pyparsing==2.4.2
+python-dateutil==2.8.0
+pytz==2019.3
+scipy==1.3.1
+six==1.12.0
