### 2019-10-16 kellykochanski/rescal-snow#41 @kellykochanski
I'm glad to hear it, and sorry that this run through wasn't as smooth as I would have liked.
