### 2019-10-16 kellykochanski/rescal-snow#41 @kellykochanski
I'm inclined to handle the sbatch issue by removing the slurm script (as my own work is all on moab, making the slurm script is harder for me to test), and encouraging users to follow their cluster's guidelines.
