### 2019-10-16 b4121e1
Removed sbatch command
