### 2019-10-16 #1699 @kellykochanski
@zbeekman If you're willing to put a little bit of time into that, I'd be delighted - I looked into CI when you first brought it up (and it does seem like a sensible solution to some of the back and forth we were having), but it's outside of my current expertise.

I'm trying to avoid relying on Jupyter notebooks, unfortunately - I do production runs in an hpc environment without access to IPython, and the tutorial is used by students learning to work in the same environment.
