### 2019-10-16 #1699 @kellykochanski
@kthyng The issues [#41](https://github.com/kellykochanski/rescal-snow/issues/41) and [#44](https://github.com/kellykochanski/rescal-snow/issues/44) are now closed, and I believe both reviewers have recommended JOSS acceptance.
