### 2019-10-16 1be5432
Responses to JOSS review (#24)
* Added toy ca code to pca directory

* Revert "Added toy ca code to pca directory"

This reverts commit 423f0b23517e1e7d1a6cf843d9ffd9414ee436cd.

Reverting accidental changes to master

* Starting to add content/outline to paper

* Fixed formatting/duplicate dir on README

* Created paper.bib, initial refs

* Create paper.bib

* Paper outlined and separated from bibliography

* Deleting corrupted LICENSE file

Correct file is in docs/LICENSE.md

* Adding correct docs/LICENSE.md file

* Correcting link to docs/LICENSE.md

* Adding license text to .py and bash fi
[…]
