+++ paper/paper.md
@@ -1,0 +1,123 @@
+---
+title: 'Rescal-snow: Simulating snow dunes with cellular automata'
+tags:
+  - C
+  - snow
+  - geomorphology
+  - Earth surface
+  - dunes
+authors:
+  - name: Kelly Kochanski
+    orcid: 0000-0003-0754-0184
+    affiliation: "1,2,3"
+  - name: Gian-Carlo Defazio
+    affiliation: "1,4"
+  - name: Eric Green
+    affiliation: 1
+  - name: Richard Barnes
+    affiliation: "5,6"
+    orcid: 0000-0002-0204-6040
+  - name: Carlos Downie
+    affiliation: 1
+  - name: Adam Rubin
+    affiliation: "7"
+  - name: Barry Rountree
+    orcid: 0000-0002-0087-4301
+    affiliation: 1
+affiliations:
+ - name: Computing Division, Lawrence Livermore National Laboratory, Livermore, CA, USA
+   index: 1
+ - name: Department of Geological Sciences, University of Colorado at Boulder, Boulder, CO, USA
+   index: 2
+ - name: Institute for Arctic and Alpine Research, University of Colorado at Boulder, Boulder, CO, USA
+   index: 3
+ - name: Department of Computer Science, Western Washington University, Bellingham, WA, USA
+   index: 4
+ - name: Energy and Resources Group, University of California Berkeley, Berkeley, CA, USA
+   index: 5
+ - name: Department of Electrical Engineering and Computer Sciences, University of California Berkeley, Berkeley, CA, USA
+   index: 6
+ - name: Department of Environmental Engineering, University of Colorado at Boulder, Boulder, CO, USA
+   index: 7
+date: 21 July 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+When wind blows over dry snow, it creates shapes known as snow bedforms.
+These features, which include snow dunes, waves, snow-steps and sastrugi (Fig. 1), ornament Antarctica, Arctic sea ice, tundra, and mountain ridges [@Filhol2015; @Kochanski2018; @Kobayashi1980].
+They change the reflectivity and average thermal conductivity of snow, and may change the patterns of snow accumulation and transport.
+Despite these effects, however, they are poorly understood and their effects are not yet included in major snow or climate models.
+
+![](../docs/example_images/field_examples.png)
+
+_Figure 1. Snow bedforms on Niwot Ridge, Colorado. From left to right, small dunes and sastrugi (looking upwind), snow dunes (looking downwind), snow-waves (looking upwind)._
+
+## Requirements of a snow bedform model
+Recent field studies have identified three new computational components needed for a good understanding of snow bedforms [@Kochanski2018; @Kochanski2019; @Filhol2015].
+First, most snow bedforms (e.g ripples, barchan dunes, snow-steps and sastrugi) are between 0.1 and 2 m in length, with select bedforms (e.g. snow-waves, some whaleback dunes) extending from 5 to 30 m.
+These length scales are based on physical phenomena such as the hop length of blowing snow grains [@Kobayashi1980] and the length scales of turbulent structures in the wind [@Kobayashi1980, @Kochanski2019].
+Existing models of wind-blown snow, however, are designed to model snow redistribution over mountainous [@Liston2007; @Lehning2002; @Marsh2018] or continental [@Gallee2012] scales
+(with grid sizes of 0.1-10 km)
+and are unable to resolve the processes that lead to the formation of snow bedforms.
+
+Second, snow bedforms are the result of a balance between sand-like granular motion and cohesive resistance [@Kochanski2019; @Filhol2015].
[…]
