### 2019-10-16 kellykochanski/rescal-snow#41 @zbeekman
@kellykochanski I'm responding over here so as to not distract the JOSS folks:

>I'm trying to avoid relying on Jupyter notebooks, unfortunately - I do production runs in an hpc environment without access to IPython, and the tutorial is used by students learning to work in the same environment.

Yes, BUT you can export the jupyter document in other formats; markdown, pdf, etc. so you could test jupyter on CI and let students follow along from markdown or rendered notebook. At any rate, if I take a stab at CI support once you've minted a release for JOSS, I'll look into this. It should be straightforward to copy the example to a notebook.
