### 2019-10-16 kellykochanski/rescal-snow#41 @kellykochanski
@zbeekman I like the idea of a regularly teted notebook that, when exported as markdown, forms something very much like the existing tutorial. I'll let JOSS go through, then move the existing tutorial into a notebook.
