### 2019-10-17 #1699 @kthyng
Paper-related changes:

- [ ] in the last paragraph of the first page, you need a semicolon between these two references: [@Kobayashi1980, @Kochanski2019]
- [ ] I think your caption for Fig 2 needs an extra line before it to get the spacing correct
- [ ] Your Liston2007 reference should have {} around {SnowTran-3D} so the capitalization is preserved.
- [ ] Same with the items in Lehning2002 reference that should be properly capitalized in your references list.
- [ ] Same for Gallee2012. You can check after you make changes by looking through the references list to make sure everything is capitalized correctly.

@openjournals/joss-eics Can anyone tell why in this paper the Marsh et al paper doesn't shorten to "Marsh et al' when referenced but instead displays all the authors? This is not consistent with other bib entries with fewer or more number of authors. And same question for the Narteau reference which displays all author names inline.
