### 2019-10-17 #1699 @kthyng
I think there are a few more bib-related errors actually; you can get more details for how to format your references [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

For example, the top two lines of reference Filhol2015 should be as below to get the order of the initial correct in the references list.

@article{Filhol2015,
  author  = "Filhol, S. and Sturm, M.",

Honestly I'm having trouble keeping track, but generally you'll want to exactly follow the examples on that site so that your references show up exactly right. In the references list, the first author should be shown as "Last name, First initial/name", but any subsequent authors should be listed as "First inital/name Last Name".
