### 2019-10-17 #1699 @kthyng
These two references have the same doi:

Narteau, C., Zhang, D., Rozier, O., & Claudin, P. (2009). Setting the length and time scales of a cellular automaton dune model from the analysis of superimposed bed forms. Journal of Geophysical Research: Earth Surface, 114(f3). doi:10.1002/esp.3479

Rozier, O., & Narteau, C. (2014). A real‐space cellular automaton laboratory. Earth Surface Processes and Landforms, 39(1), 98–109. doi:10.1002/esp.3479
