### 2019-10-18 3807811
Typos + capitalizations in paper.bib
Response to editorial comments by kthyng
