+++ paper/paper.bib
@@ -1,13 +1,3 @@
-@article{Clark2012,
-  author  = "Clark, M. P., and J. Hendrikx, and A. G. Slater, and D. Kavetski, and B. Anderson, and N. J. Cullen, and T. Kerr, and E. O Hreinsson, and R. A. Woods",
-  year    = 2011,
-  title   = "Representing spatial variability of snow water equivalent in hydrologic and land-surface models: A review",
-  journal = "Water Resources Research",
-  volume  = 47,
-  issue   = W07539,
-  doi     = "10.1029/2011WR010745"
-}
-
 @article{Narteau2014,
   author  = "Rozier, O., and Narteau, C.",
   year    = 2014,
@@ -26,38 +16,18 @@
   journal = "Journal of Geophysical Research: Earth Surface",
   volume  = 114,
   issue   = F3,
-  doi     = "10.1002/esp.3479"
+  doi     = "10.1029/2008JF001127"
 }
 
 @article{Liston2007,
-  author  = "Liston, G. E., and Haehnel, R. B., and Sturm, M., and Hiemstra, and C. A., Berezovskaya, S., and Tabler, R. D.",
+  author  = "Liston, G. E., and Haehnel, R. B., and Sturm, M., and Hiemstra, C. A., and Berezovskaya, S., and Tabler, R. D.",
   year    = 2007,
-  title   = "Simulating complex snow distributions in windy environments using SnowTran-3D",
+  title   = "Simulating complex snow distributions in windy environments using {SnowTran-3D}",
   journal = "Journal of Glaciology",
   volume  = 53,
   issue   = 181, 
   pages   = "241-256",
   doi     = "10.3189/172756507782202865"
-}
-
-@article{Lenaerts2012,
-  author  = "Lenaerts, J. T. M. and van den Broeke, M. R. and  van de Berg, W. and van Meijgaard, E., and Munneke, P. K.",
-  title   = "A new, high-resolution surface mass balance map of Antarctica (1979-2010) based on regional atmospheric climate modeling",
-  journal = "Geophysical Research Letters",
-  volume  = 39,
-  issue   = 4,
-  year    = 2012,
-  doi     = "10.1029/2011GL050713"
-}
-
-@article{Blanchard2015,
-  author  = "Blanchard-Wrigglesworth, and E., S. L. Farrell, and T. Newman, and C. M. Bitz",
-  title   = "Snow cover on Arctic sea ice in observations and an Earth System Model",
-  journal = "Geophysical Research Letters",
-  volume  = 42,
-  pages   = "10342--10348",
-  doi     = "10.1002/2015GL066049",
-  year    = 2015
 }
 
 @article{Colbeck1983,
@@ -73,7 +43,7 @@
 
 @article{Lehning2002,
   author  = "Lehning, M. and Bartelt, P. and  Brown, B. and Fierz, C.",
-  title   = "A physical SNOWPACK model for the Swiss avalanche warning: Part III: meteorological forcing, thin layer formation and evaluation",
+  title   = "A physical {SNOWPACK} model for the Swiss avalanche warning: {Part III}: meteorological forcing, thin layer formation and evaluation",
   journal = "Cold Regions Science and Technology",
   volume  = 35,
   issue   = 3,
@@ -89,17 +59,18 @@
   journal = "Journal of Geophysical Research: Earth Surface",
   volume  = 120, 
   issue   = 9, 
+  pages   = "1645--1669"
   doi     = "10.1002/2015JF003529"
 }
 
 @article{Gallee2012,
-  author  = "Gall\'ee, H., and A. Trouvilliez, and C. Agosta, and C. Genthon, and V. Favier, and F. Naiim-Bouvet",
[…]
