### 2019-10-18 7ebd620
Correct paper.md, paper.bib in response to editorial comments (#47)
* Polishing paper.md

* Typos + capitalizations in paper.bib
