+++ paper/paper.bib
@@ -59,7 +59,7 @@
   journal = "Journal of Geophysical Research: Earth Surface",
   volume  = 120, 
   issue   = 9, 
-  pages   = "1645--1669"
+  pages   = "1645--1669",
   doi     = "10.1002/2015JF003529"
 }
 
