### 2019-10-18 5c05b68
Umlaut for Lammel in paper.bib (#48)
