+++ paper/paper.bib
@@ -94,7 +94,7 @@
 }
 
 @article{Lammel2012,
-  author  = "L\:ammel, M., and Rings, D. and Kroy, K.",
+  author  = "L&uuml;ammel, M., and Rings, D. and Kroy, K.",
   year    = "2012",
   title   = "A two-species continuum model for aeolian sand transport",
   journal = "New Journal of Physics",
