+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{Narteau2009,
-  author  = "Narteau, C., and Zhang, D., and Rozier, O., and Claudin, P.",
+  author  = {{Narteau}, C., and {Zhang}, D., and {Rozier}, O., and {Claudin}, P.},
   year    = 2009,
   title   = "Setting the length and time scales of a cellular automaton dune model from the analysis of superimposed bed forms",
   journal = "Journal of Geophysical Research: Earth Surface",
@@ -20,7 +20,7 @@
 }
 
 @article{Liston2007,
-  author  = "Liston, G. E., and Haehnel, R. B., and Sturm, M., and Hiemstra, C. A., and Berezovskaya, S., and Tabler, R. D.",
+  author  = {Liston, G. E., and {Haehnel}, R.~B., and {Sturm}, M., and {Hiemstra}, C.~A., and {Berezovskaya}, S., and {Tabler}, R.~D.},
   year    = 2007,
   title   = "Simulating complex snow distributions in windy environments using {SnowTran-3D}",
   journal = "Journal of Glaciology",
@@ -42,7 +42,7 @@
 }
 
 @article{Lehning2002,
-  author  = "Lehning, M. and Bartelt, P. and  Brown, B. and Fierz, C.",
+  author  = {{Lehning}, M. and {Bartelt}, P. and  {Brown}, B. and {Fierz}, C.},
   title   = "A physical {SNOWPACK} model for the Swiss avalanche warning: {Part III}: meteorological forcing, thin layer formation and evaluation",
   journal = "Cold Regions Science and Technology",
   volume  = 35,
@@ -64,7 +64,7 @@
 }
 
 @article{Gallee2012,
-  author  = "Gall\'ee, H., and Trouvilliez, A. and Agosta, C. and Genthon, C. and Favier, V. and Naiim-Bouvet, F.",
+  author  = {{Gall\'ee}, H., and {Trouvilliez}, A. and {Agosta}, C. and {Genthon}, C. and {Favier}, V. and {Naiim-Bouvet}, F.},
   year    = 2012, 
   title   = "Transport of snow by the wind: A comparison between observations in {A}d\'elie {L}and, {Antarctica}, and simulations made with the regional climate model {MAR}", 
   journal = "Boundary-Layer Meteorology", 
@@ -84,7 +84,7 @@
 }
 
 @article{Kochanski2019,
-  author  = "Kochanski, K., and Anderson, R. S. and Tucker, G. E.",
+  author  = {{Kochanski}, K., and {Anderson}, R.~S. and {Tucker}, G.~E.",
   year    = 2019, 
   title   = "The evolution of snow bedforms in the {Colorado Front Range} and the processes that shape them", 
   journal = "The Cryosphere", 
@@ -94,7 +94,7 @@
 }
 
 @article{Lammel2012,
-  author  = "L&uuml;ammel, M., and Rings, D. and Kroy, K.",
+  author  = {{Lämmel}, M., and {Rings}, D. and {Kroy}, K.},
   year    = "2012",
   title   = "A two-species continuum model for aeolian sand transport",
   journal = "New Journal of Physics",
@@ -104,7 +104,7 @@
 }
 
 @article{Marsh2018,
-  author  = "Marsh, C. B. and Spiteri, R. J. and Pomeroy, J. W. and Wheater, H. S.",
+  author  = {{Marsh}, C.~B. and {Spiteri}, R.~J. and {Pomeroy}, J.~W. and {Wheater}, H.~S.},
   year    = "2018",
   journal = "Computers & Geosciences",
   title   = "Multi-objective unstructured triangular mesh generation for use in hydrological and land surface models",
[…]
