+++ paper/paper.bib
@@ -84,7 +84,7 @@
 }
 
 @article{Kochanski2019,
-  author  = {{Kochanski}, K., and {Anderson}, R.~S. and {Tucker}, G.~E.",
+  author  = {{Kochanski}, K., and {Anderson}, R.~S. and {Tucker}, G.~E.},
   year    = 2019, 
   title   = "The evolution of snow bedforms in the {Colorado Front Range} and the processes that shape them", 
   journal = "The Cryosphere", 
