### 2019-10-18 #1699 @kellykochanski
@kthyng I've fixed the following:

- Figure 2 caption has its own line
- Missing semicolon added to bibliography
- Missing umlaut in "Lammel" name properly formatted
- Capitalizations in paper names {bracketed}
- DOIs checked

I have not been able to fix the expanded author list for Marsh et al., despite reformatting (from quotes "" to brackets {{}}) to match the example you sent.

Currently, lists of 1-4 authors are shown in full, and the papers with 5+ authors are abbreviated to Firstauthor et al.
