### 2019-10-22 #1699 @kthyng
Nice, thanks @kellykochanski. We can proceed.
