### 2019-10-22 c8e91ad
Adding major updates to NEWS
