### 2019-10-23 #1699 @kellykochanski
@kthyng I've created a software arxiv on zenodo with doi 10.5281/zenodo.3517217

dx.doi.org/10.5281/zenodo.3517217
